% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathy_match}
\alias{bathy_match}
\title{Extract depth for position from GEBCO}
\usage{
bathy_match(x, y)
}
\arguments{
\item{x}{latitude in decimal degrees}

\item{y}{longitude in decimal degrees}
}
\value{
depth in meters
}
\description{
extract nearest point from GEBCO2019
}
