% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogeochem.R
\name{biogeochem.fetch}
\alias{biogeochem.fetch}
\title{Fetch data from the biogeochemical database}
\usage{
biogeochem.fetch(
  survey = NA,
  bbox = NA,
  after = NA,
  before = NA,
  parameters = c("SAL", "TOxN")
)
}
\arguments{
\item{survey}{a survey code, use '\%' as a wildcard e.g. "CEND_\%_18"}

\item{bbox}{a named vector bounding box (ymin, ymax, xmin, xmax) as generated by calc_bounding_box}

\item{after}{a time stamp after which data is selected e.g. "2018-01-01"}

\item{before}{a time stamp before which data is selected e.g. "2019-01-01"}

\item{parameters}{a vector of parameter codes default is c("SAL", "TOXN"), use parameters = "ALL" to select all.}
}
\value{
data.table
}
\description{
Tool to query the "data" view from the biogeochemical postgres database
}
\details{
These data use the following flag scheme
\tabular{rr}{
\strong{flag} \tab \strong{flag_description} \cr
0 \tab         no quality control \cr
1 \tab                 good value \cr
2 \tab        probably good value \cr
3 \tab         probably bad value \cr
4 \tab                  bad value \cr
5 \tab              changed value \cr
6 \tab      value below detection \cr
7 \tab            value in excess \cr
8 \tab         interpolated value \cr
9 \tab              missing value \cr
A \tab value phenomenon uncertain \cr
B \tab              nominal value
}
}
\examples{
\dontrun{
bbox = calc_bounding_box(52.2, 2, 500) # 1 km box centred on 52.2N 2E
dat = biogeochem.fetch(bbox = bbox)
dat = biogeochem.fetch(after = "2018-01-01", parameters = c("all"))
}
}
