% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{calc_bounding_box}
\alias{calc_bounding_box}
\title{Calculate bounding box}
\usage{
calc_bounding_box(lat, lon, size = 250)
}
\arguments{
\item{lat}{vector of lat}

\item{lon}{vector of lon}

\item{size}{distance in meters}
}
\value{
vector of limits (north, east, south, west)
}
\description{
used WGS84 elipsoid to calculate northern, eastern, southern and western extent from a starting lat/lon.
}
\examples{
calc_bounding_box(51, 1.2, 500) # 1x1km square
}

