% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{calc_sal}
\alias{calc_sal}
\title{Calculate salinity}
\usage{
calc_sal(Cond, t, p = max(0, P - 1.013253), P = 1.013253)
}
\arguments{
\item{Cond}{}

\item{t}{}

\item{p}{}

\item{P}{}
}
\value{
salinity
}
\description{
Calculate salinity
}

