% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cds_table}
\alias{cds_table}
\title{cds_table}
\format{
A List of 4095
}
\usage{
cds_table
}
\description{
a precompiled list of all possible combinations of Cefas Endeavour ferrybox error codes
This is used as a lookup table
}
\examples{
cds_table[3] # underflow and value too low
}
\keyword{datasets}
