% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{convert.clearwater}
\alias{convert.clearwater}
\title{Clearwater data converter and QA1}
\usage{
convert.clearwater(file = NA, parcode = "TOXN", return_excel = TRUE)
}
\arguments{
\item{file}{csv file from Clearview software, if left blank will open dialog.}

\item{parcode}{string matching the SmartBuoy database parameter code, currently only "TOXN"}

\item{return_excel}{if TRUE (default) writes an excel file, otherwise returns data.table.

This tool converts the output from the Clearwater software,
applies a simple in-out of range check to the data (QA level 1) and returns a data.table or optionally a excel file ready for upload
the ranges applied are:

1, Long 80mm (range 0.1-20 uM)
2, Medium 10 mm (range < 100 uM)
3, Short 1mm (range < 1000 uM)

The function then selects the best range.}
}
\value{
Either a data.table or a string if a excel file is written
}
\description{
Clearwater data converter and QA1
}
\examples{
convert.clearwater("WGAB2_CWN54/WESTGAB2_000_CWN54.csv")
}
