% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geo.R
\name{convert_latlong}
\alias{convert_latlong}
\title{Convert degrees + decimal minutes to decimal degrees}
\usage{
convert_latlong(degrees, decimal_minutes)
}
\arguments{
\item{degrees}{}

\item{decimal_minutes}{}
}
\value{
decimal degrees
}
\description{
Convert degrees + decimal minutes to decimal degrees
}

