% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{convert_latlong_ddmmm}
\alias{convert_latlong_ddmmm}
\title{Convert decimal degrees  to degrees + decimal minutes}
\usage{
convert_latlong_ddmmm(degrees, paste = T)
}
\arguments{
\item{degrees}{numeric vector of decimal degrees}

\item{paste}{if True (default) return formatted string, otherwise returns list of DD and MM}
}
\value{
decimal degrees
}
\description{
Convert decimal degrees  to degrees + decimal minutes
}
