% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{convert_latlong_ddmmss}
\alias{convert_latlong_ddmmss}
\title{Convert degrees + minutes + seconds to decimal degrees}
\usage{
convert_latlong_ddmmss(degrees, minutes, seconds, polarity = NA)
}
\arguments{
\item{degrees}{numeric vector of whole degrees, if negative will convert to W or S.}

\item{minutes}{numeric vector of minutes}

\item{seconds}{numeric vector of seconds}

\item{polarity}{optional "E, W, N, S"}
}
\value{
decimal degrees
}
\description{
Convert degrees + minutes + seconds to decimal degrees
}
