% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{ctd.fetch}
\alias{ctd.fetch}
\title{Fetch CTD data from CTD postgres database}
\usage{
ctd.fetch(
  survey = NA,
  bbox = NA,
  after = NA,
  before = NA,
  parameters = c("TEMP", "SAL", "O2CONC", "FTU", "FLUORS", "ALT", "PAR"),
  binned = T
)
}
\arguments{
\item{survey}{a survey code, use '\%' as a wildcard e.g. "\%CEND\%18"}

\item{bbox}{a named vector bounding box (ymin, ymax, xmin, xmax) as generated by calc_bounding_box}

\item{after}{a time stamp after which data is selected e.g. "2018-01-01"}

\item{before}{a time stamp before which data is selected e.g. "2019-01-01"}

\item{parameters}{a vector of parameter codes default is c("TEMP", "SAL", "O2CONC", "FTU", "FLUORS", "ALT", "PAR"), use parameters = "ALL" to select all.}

\item{binned}{if TRUE (default) return 1 dbar binned data, otherwise returns raw frequency data}
}
\value{
data.table containing binned CTD data
}
\description{
Fetch CTD data from CTD postgres database
}
\examples{
\dontrun{
bbox = calc_bounding_box(52.2, 2, 500) # 1 km box centred on 52.2N 2E
dat = ctd.fetch(bbox = bbox)
dat = ctd.fetch(after = "2018-01-01", parameters = c("SAL", "V1", "ALT"))
}
}
