% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dircast}
\alias{dircast}
\title{mark up and down casts}
\usage{
dircast(pressure)
}
\arguments{
\item{pressure}{a order vector of pressures, any unit}
}
\value{
vector of same length as pressure, with -1 for descending and 1 for ascending (after deepest point)
}
\description{
finds the deepest point of pressure record
}
