% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{dTmatch}
\alias{dTmatch}
\title{Fast fuzzy dateTime matcher
j}
\usage{
dTmatch(dat, reference, varname, max_dT = Inf)
}
\arguments{
\item{dat}{data.table of variables you want to look up, must contain a POSIXct "dateTime" column}

\item{reference}{data.table of variable you want to search against, must contain a POSIXct "dateTime" column}

\item{varname}{column name in `reference` you want to add to `dat`}

\item{max_dT}{integer value in seconds for maximum permissible gap between lookup value}
}
\value{
dat data.table with added variable column and dateTimes from `reference`
}
\description{
Fast fuzzy dateTime matcher
j
}
\details{
finded closest timestamp in `reference`.
}

