% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnKF.R
\name{EnKF}
\alias{EnKF}
\title{Ensemble Kalman Filter (EnKF)}
\usage{
EnKF(
  y,
  mod,
  GGfunction,
  FFfunction,
  size = 50,
  logLik = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{y}{data. The data can be a vector (one-dimensional measurements / # univariate time series)
or a matrix (multidimensional measurements /
multivariate time series). In the matrix case, each of the columns contains
the measurements on a single dimension (or, similarly, a single \link{univariate}
time series).
Missing values (NA's) are allowed.}

\item{mod}{a list with the components m0 (initial state estimates),C0 (error covariances of the initial state estimates), V (measurement noise), and W (process noise).}

\item{GGfunction}{a function with arguments x and k. The GGfunction specifies the
state transition function. The transition function is usually denoted as
f(x\link{k}, u\link{k}), where x are the states estimates and u the control input at
time step k.}

\item{FFfunction}{a function with arguments x and k. The FFfunction specifies the
observation/measurement function. The measurement function is usually denoted as h(x\link{k}).}

\item{size}{size is a integer specifying the ensemble size.}

\item{logLik}{if TRUE, then return the negative loglikelihood of the specified model.}

\item{simplify}{if TRUE, then do not include the data (=y) in the output.}
}
\value{

}
\description{
This function implements the Ensemble Kalman Filter (EnKF) as described
by Gillijns and colleagues in their 2006 paper "What Is the Ensemble Kalman #' Filter and How Well Does it Work?".
Note that this EnKF assumes (time invariant) additive noise.
}
\author{
Stefan Gelissen (info@datall-analyse.nl)
}
