% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnKF.R
\name{EnKFforecast}
\alias{EnKFforecast}
\title{EnKF forecast}
\usage{
EnKFforecast(filterData, nAhead = 1)
}
\value{

}
\description{
predicts future system states and observations for the Ensemble Kalman Filter (EnKF).
}
\author{
Stefan Gelissen (info@datall-analyse.nl)
}
