% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expanded_range}
\alias{expanded_range}
\title{Expanded range calculation}
\usage{
expanded_range(x, scale = 1.2)
}
\arguments{
\item{x}{a vector}

\item{scale}{factor by which to scale the range, default is 1.2 (20 \% increase)}
}
\value{
two element vector of min and max
}
\description{
Calculate the range of a vector, and then scale it. Useful when calculating xlims for plotting.
}
\examples{
lons = c(-9, 5.5, 9)
expanded_range(lons)
}
