% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fdiff}
\alias{fdiff}
\title{First order differentiation}
\usage{
fdiff(x, t = NA)
}
\arguments{
\item{x}{vector of values to differentiate}

\item{t}{optional time in seconds, or POSIXct}
}
\description{
This is an implementation of Sunke Schmidtko's method for a first order differentiation of a vector.
Unlike the standard R diff function this returns a vector of the same length, coping with end effects.
If time is supplied to the optional "t" variable then the value returned will be dx/dt.
Otherwise it will just be dx.
}
