% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.conlog}
\alias{ferrybox.conlog}
\title{Ferrybox conlog data reader}
\usage{
ferrybox.conlog(conlog)
}
\arguments{
\item{conlog}{character string indicating device data file}
}
\value{
data.frame
}
\description{
reads DAQ server conlog data files
}
\details{
TODO
requires intensive string operations on many rows and thus fairly slow
}
\keyword{conlog}
\keyword{ferrybox}

