% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.course}
\alias{ferrybox.course}
\title{Calculate course from GPS}
\usage{
ferrybox.course(dateTime, lat, lon, threshold = 65)
}
\arguments{
\item{dateTime}{in POSIXct}

\item{lat}{in decimal degrees}

\item{threshold}{maximum time interval between gps points (seconds)}
}
\value{
vector of courses
}
\description{
Calculate course from GPS
}

