% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.cruiselist}
\alias{ferrybox.cruiselist}
\title{ferrybox cruise id queryer}
\usage{
ferrybox.cruiselist(yr = "ALL", db_name = "ferrybox")
}
\arguments{
\item{yr}{integer specifing a year to limit the search, default is 'ALL'}

\item{db_name}{character string matching ODBC data source name, defaults to 'ferrybox'}
}
\value{
character vector of Cruise Id's
}
\description{
Fetches lists of cruise id
}
\details{
This function querys the Smartbuoy database and returns a list of valid cruise IDs matching the supplied critiera where ESM2
profiler data is available. The v_CtdProfile_AllData table is used, as such private data will not be available to this function.
}
\keyword{ferrybox}
\keyword{query}

