% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.devdata}
\alias{ferrybox.devdata}
\title{Ferrybox device data reader}
\usage{
ferrybox.devdata(devdata_folder, pivot = F)
}
\arguments{
\item{devdata_folder}{character string indicating path to folder containing device data files}

\item{pivot}{optional boolian indicating if returned table should be recast into 'wide' format}
}
\value{
data.frame
}
\description{
reads ferrybox device data files
}
\details{
TODO
}
\keyword{ferrybox}

