% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.errorcode}
\alias{ferrybox.errorcode}
\title{Ferrybox error codes}
\usage{
ferrybox.errorcode(x)
}
\arguments{
\item{x}{error code}
}
\value{
codes used in composite
}
\description{
Ferrybox error codes
}

