% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.fetch}
\alias{ferrybox.fetch}
\title{Ferrybox query tool}
\usage{
ferrybox.fetch(
  cruiseID = NA,
  after = NA,
  before = NA,
  area = NA,
  parameters = c("TEMP", "SAL", "FTU", "O2CONC"),
  min_QA_reached = TRUE,
  most_recent = NA,
  RQ0 = TRUE,
  hull_temp_only = FALSE,
  db_name = "ferrybox"
)
}
\arguments{
\item{cruiseID}{optional cruise ID string, e.g. "CEND_02_14", if provided only data from this cruise will be returned.}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{area}{optional vector consisting of 4 elements, minimum Latitude, minimum Longitude, maximum Latitude, maximum Longitude e.g. c(52, -3.5, 53, -4)}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC')}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{RQ0}{boolean, if True only data where result quality = 0 is returned, i.e. good data, default it TRUE}

\item{hull_temp_only}{boolean, if True only temperature data from the hull temperature probe will be returned}

\item{db_name}{character string matching ODBC data source name, defaults to 'ferrybox'}
}
\value{
data.frame with returned data in "long" format or error string if no data returned
}
\description{
Fetches ferrybox data
}
\details{
TODO
}
\keyword{ferrybox}
\keyword{query}
