% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.heading}
\alias{ferrybox.heading}
\title{Calculate heading from GPS}
\usage{
ferrybox.heading(dateTime, lat, lon, threshold = 120)
}
\arguments{
\item{dateTime}{in POSIXct}

\item{lat}{in decimal degrees}

\item{threshold}{maximum time interval between gps points (seconds)}
}
\value{
vector of headings
}
\description{
Calculate heading from GPS
}

