% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.position}
\alias{ferrybox.position}
\title{ferrybox cruise tracks}
\usage{
ferrybox.position(cruiseID = NA, after = NA, before = NA,
  db_name = "ferrybox")
}
\arguments{
\item{cruiseID}{optional character string matching cruise ID}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{db_name}{character string matching ODBC data source name, defaults to 'ferrybox'}
}
\value{
data.frame (data.table) containing the ferrybox cruise track
}
\description{
Fetches position of ferrybox
}
\details{
TODO
}
\keyword{ferrybox}
\keyword{query}

