% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.read_port_xml}
\alias{ferrybox.read_port_xml}
\title{Read ferrybox port positions xml file}
\usage{
ferrybox.read_port_xml(xmlfile = "data/Positions.xml")
}
\arguments{
\item{xmlfile}{}
}
\value{
data.table of positions and watch circles
}
\description{
Read ferrybox port positions xml file
}
