% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.speed}
\alias{ferrybox.speed}
\title{calculate speed from GPS}
\usage{
ferrybox.speed(dateTime, lat, lon, threshold = 65)
}
\arguments{
\item{dateTime}{in POSIXct}

\item{lat}{in decimal degrees}

\item{threshold}{maximum time interval between gps points (seconds)}
}
\value{
vector of speeds
}
\description{
calculate speed from GPS
}

