% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{fetch.ferrybox}
\alias{fetch.ferrybox}
\title{Ferrybox query tool}
\usage{
fetch.ferrybox(cruiseID = NA, after = NA, before = NA, area = NA,
  parameters = c("TEMP", "SAL", "FTU", "O2CONC"), RQ0 = TRUE,
  db.name = "ferrybox")
}
\arguments{
\item{optional}{cruise ID string, e.g. "CEND_02_14", if provided only data from this cruise will be returned.}

\item{optional}{date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{optional}{date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{optional}{area vector consisting of 4 elements, minimum Latitude, minimum Longitude, maximum Latitude, maximum Longitude e.g. c(52, -3.5, 53, -4)}

\item{optional}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC')}

\item{optional}{boolean, if True only data where result quality = 0 is returned, i.e. good data}

\item{optional}{character string matching ODBC data source name, defaults to 'ferrybox'}
}
\value{
data.frame with returned data in "long" format or error string if no data returned
}
\description{
Fetches ferrybox data
}
\details{
TODO
}
\examples{
d <- fetch.ferrybox('CEND_01_14')
}
\keyword{ferrybox}
\keyword{query}

