% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{fetch.smartbuoy}
\alias{fetch.smartbuoy}
\title{SmartBuoy data query tool}
\usage{
fetch.smartbuoy(deployment = NA, deployment_group = NA, after = NA,
  before = NA, parameters = c("TEMP", "SAL", "FTU", "O2CONC", "PAR"),
  min_QA_reached = TRUE, RQ0 = TRUE, ct_temp_only = TRUE,
  db_name = "smartbuoydblive")
}
\arguments{
\item{deployment}{optional string matching smartbuoy deployment}

\item{deployment_group}{optional string matching smartbuoy deployment_group}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC', 'PAR')}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{RQ0}{boolean, if True only data where result quality = 0 is returned, i.e. good data. default is True}

\item{ct_temp_only}{boolean, if True all non-FSI temperature data is discarded, default is False.}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table with returned data in "long" format or error string if no data returned
}
\description{
Fetches SmartBuoy data
}
\details{
This function querys the Smartbuoy database and returns
}
\keyword{esm2}
\keyword{query}
\keyword{smartbuoy}

