% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{findMLD}
\alias{findMLD}
\title{Find mixed layer depth}
\usage{
findMLD(d, p, threshold = 0.125)
}
\arguments{
\item{d}{}

\item{p}{}

\item{threshold}{}
}
\value{
mld
}
\description{
Find mixed layer depth
}

