% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{ftu_from_ADC}
\alias{ftu_from_ADC}
\title{ftu from ADC}
\usage{
ftu_from_ADC(x, factor = 1.22, offset = 0.001)
}
\arguments{
\item{x}{vector of ESM2 hex ADC counts, each count should be 6 characters i.e. "1a13D2"}

\item{factor}{channel calibration factor, specific to logger, default is 1.22}

\item{offset}{channel calibration offset, default is 0.001}
}
\value{
vector of calibrated FTU values
}
\description{
Calculates FTU value from ESM2 raw hex ADC counts
}
\details{
This function querys the Smartbuoy database and returns
}
\keyword{esm2}

