% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{fuzzy_spacetime}
\alias{fuzzy_spacetime}
\title{Improved Spatial-temporal matching}
\usage{
fuzzy_spacetime(A, Z, tt = 3600, dt = 5000)
}
\arguments{
\item{A}{data.table with lat, lon and dateTime columns}

\item{Z}{data.table with lat, lon and dateTime columns}

\item{tt}{time in seconds to match within}

\item{dt}{distance in meters to match within}

\item{merge}{bool}
}
\value{
if merge is true (default) returns single combined data.table of matching values, else returns list of the two subset data.tables
}
\description{
builds distance matrix between all points in two data.tables using WGS84 elipsoid,
this matrix is then reduced to those within a specified distance threshold.
These spacial matched points are then matched against time within a time threshold
}
