% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fuzzymatch}
\alias{fuzzymatch}
\title{Fast fuzzy matcher}
\usage{
fuzzymatch(dat, reference, index = "dateTime", threshold = Inf, return = F)
}
\arguments{
\item{dat}{data.table of variables you want to look up, typically contain a POSIXct "dateTime" index column.}

\item{reference}{data.table of variable you want to search against, must contain the same named index column.}

\item{index}{column name in \code{reference} you want to match against (typically "dateTime")}

\item{threshold}{integer value for maximum permissible gap between lookup value, unit = seconds for "dateTime".}

\item{return}{if false (default) values with no match, or outside threshold are dropped}
}
\value{
data.table with added variable column and index from \code{reference}
}
\description{
Fast fuzzy matcher
}
\details{
finded closest index in \code{reference}.

Matches values from a reference data.table to another. Typically used for fuzzy dateTime matching.
note, make sure the index column is the same data type in both dat and reference. e.g. as.numeric
}
\examples{
dat = data.table(i = c(4.3, 5.9, 1.2), datval = runif(3)+10, datstuff="test")
reference = data.table(i = as.numeric(1:10), refjunk = "junk", refval = runif(10))
fuzzymatch(dat, reference, index="i")

}
