% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gebco_2019}
\alias{gebco_2019}
\title{GEBCO 2019}
\format{
A data frame with 3693762 rows and 3 variables:
\describe{
\item{\code{lon}}{WGS84 Longitude in decimal degrees}
\item{\code{lat}}{WGS84 Latitude in decimal degrees}
\item{\code{depth}}{depth in meters}
}
}
\source{
\url{https://www.gebco.net/}
}
\usage{
gebco_2019
}
\description{
GEBCO bathymetry at 0.01 degree grid centred on the UK (45-65N -16-10E).
}
\keyword{datasets}
