% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gebco_low}
\alias{gebco_low}
\title{GEBCO Low Resolution}
\format{
A data table
\describe{
\item{\code{lon}}{WGS84 Longitude in decimal degrees}
\item{\code{lat}}{WGS84 Latitude in decimal degrees}
\item{\code{depth}}{depth in meters}
\item{\code{label}}{label corisponding to depth bin}
}
}
\source{
\url{https://www.gebco.net/}
}
\usage{
gebco_low
}
\description{
GEBCO 2024 bathymetry at 0.05 degree grid centred on the UK (44-65N -17-11E).
}
\keyword{datasets}
