% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ggmap.fetch}
\alias{ggmap.fetch}
\title{fetch localised ggmap}
\usage{
ggmap.fetch(lat, lon, zoom_to_group = T, scale_factor = 0, crop = F,
  maptype = "satellite")
}
\arguments{
\item{lat}{optional vector of decimal latitudes}

\item{lon}{optional vector of decimal longitudes}

\item{zoom_to_group}{boolean, if True map is centred and zoomed to input lat/long, if False entire UK is used.}

\item{scale_factor}{optional integer, increase set to 1 (or more) to pad to next zoom level}

\item{crop}{boolean, if True map is cropped to lat lon + crop_padding (default is False)}

\item{maptype}{string of either "satellite", "terrain", "terrain-background", "hybrid"}
}
\value{
ggmap object
}
\description{
Creates a base map centred and scaled
}
\details{
TODO
}
\keyword{map}

