% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggwavelet}
\alias{ggwavelet}
\title{ggplot biwavelet}
\usage{
ggwavelet(wt, base = 2, colors = "viridis", isPOSIXct = T,
  yscale = "hours", ylim = NA)
}
\arguments{
\item{wt}{a wavelet object created with biwavelet::wt}

\item{base}{log scale to use}

\item{colors}{either "viridis" (default), "inferno" or "jet"}

\item{isPOSIXct}{if True (default), assume time is in seconds}

\item{yscale}{if ifPOSIXct is True set this to "hours" (default), "seconds" or "days"}

\item{ylim}{optional, two element vector containing limits of y in units of yscale}
}
\value{
ggplot
}
\description{
Plots a biwavelet wt object with sensible defaults
}

