% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.cruiselist}
\alias{lims.cruiselist}
\title{Fetch list of cruises from LIMS}
\usage{
lims.cruiselist(yr = "ALL", db_name = "lims")
}
\arguments{
\item{yr}{integer specifing a year to limit the search, default is 'ALL'}

\item{db_name}{character string matching ODBC data source name, defaults to 'lims'}
}
\value{
vector of cruise ids
}
\description{
Fetch list of cruises from LIMS
}
