% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.fetch.historic}
\alias{lims.fetch.historic}
\title{Fetch nutrients data from LIMS (pre 2013)}
\usage{
lims.fetch.historic(parameters = c("SAL", "CHLOROPHYLL", "SLD", "TOXN", "O2"),
  db_name = "lims")
}
\arguments{
\item{db_name}{}
}
\value{
data frame (data.table) of extracted nutrients data
}
\description{
Fetch nutrients data from LIMS (pre 2013)
}

