% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapdata}
\alias{mapdata}
\title{Western European shelf coastlines}
\format{
A data frame with 50873 rows and 7 variables:
\describe{
\item{\code{lon}}{WGS84 longitude in decimal degees}
\item{\code{lat}}{WGS84 latitude in decimal degees}
\item{\code{id}}{country name}
\item{\code{group}}{polygon group}
}
}
\usage{
mapdata
}
\description{
a subset from rworldmap high resolution, but just for the Western European shelf countries
}
\references{
\code{rworldmap}
}
\keyword{datasets}
