% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{markday}
\alias{markday}
\title{Calculate day or night}
\usage{
markday(dateTime, lat, lon)
}
\arguments{
\item{lon}{}
}
\value{
vector of "day" or "night" character values
}
\description{
Calculate day or night
}

