% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{match_spacetime}
\alias{match_spacetime}
\title{Spatial-temporal matching}
\usage{
match_spacetime(
  x,
  y,
  distance_threshold = 5000,
  time_threshold = 3600,
  merge = T
)
}
\arguments{
\item{x}{data.table with lat, lon and dateTime columns}

\item{y}{data.table with lat, lon and dateTime columns}

\item{distance_threshold}{distance in meters to match within}

\item{time_threshold}{time in seconds to match within}

\item{merge}{bool}
}
\value{
if merge is true (default) returns single combined data.table of matching values, else returns list of the two subset data.tables
}
\description{
builds distance matrix between all points in two data.tables using WGS84 elipsoid,
this matrix is then reduced to those within a specified distance threshold.
These spacial matched points are then matched against time within a time threshold
}
