% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{nc_print}
\alias{nc_print}
\title{print the variables in a netcdf (ncdump lite)}
\usage{
nc_print(ncfile)
}
\arguments{
\item{ncfile}{a netcdf filename}
}
\value{
variable names
}
\description{
print the variables in a netcdf (ncdump lite)
}
