% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen (MSI's conflicted copy 2019-12-16).R,
%   R/oxygen.R
\name{optode.analogCalphase}
\alias{optode.analogCalphase}
\title{Analog optode calphase}
\usage{
optode.analogCalphase(v, PhaseLimit0 = 10, PhaseLimit1 = 70)

optode.analogCalphase(v, PhaseLimit0 = 10, PhaseLimit1 = 70)
}
\arguments{
\item{v}{volts}

\item{TempLimit0}{optode PhaseLimit lower parameter (usually 10)}

\item{TempLimit1}{optode PhaseLimit upper parameter (usually 70)}

\item{v}{volts}

\item{TempLimit0}{optode PhaseLimit lower parameter (usually 10)}

\item{TempLimit1}{optode PhaseLimit upper parameter (usually 70)}
}
\value{
vector of temperature (C)

vector of temperature (C)
}
\description{
Calculates Calphase from optode analog channel

Calculates Calphase from optode analog channel
}
