% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen (MSI's conflicted copy 2019-12-16).R,
%   R/oxygen.R
\name{optode.analogtemp}
\alias{optode.analogtemp}
\title{Analog optode temperature}
\usage{
optode.analogtemp(v, TempLimit0 = -5, TempLimit1 = 35)

optode.analogtemp(v, TempLimit0 = -5, TempLimit1 = 35)
}
\arguments{
\item{v}{volts}

\item{TempLimit0}{optode TempLimit lower parameter (usually -5)}

\item{TempLimit1}{optode TempLimit upper parameter (usually 35)}

\item{v}{volts}

\item{TempLimit0}{optode TempLimit lower parameter (usually -5)}

\item{TempLimit1}{optode TempLimit upper parameter (usually 35)}
}
\value{
vector of temperature (C)

vector of temperature (C)
}
\description{
Calculates temperature from optode analog channel

Calculates temperature from optode analog channel
}
