% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optode.R
\name{optode.correction}
\alias{optode.correction}
\title{Optode salinity and pressure compensation}
\usage{
optode.correction(O2, t, S, depth = 0, optode_salinity = 0)
}
\arguments{
\item{O2}{oxygen concentration in mmol m-3, (umol / l) or ml/l}

\item{t}{temperature}

\item{S}{salinity}

\item{depth}{depth of sensor in meters}

\item{optode_salinity}{internal optode setting default is 0}
}
\value{
corrected oxygen concentration in units of `O2`
}
\description{
Optode salinity and pressure compensation
}
\examples{
optode.correction(300, 7.701, 34.176, 1000)
}

