% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen (MSI's conflicted copy 2019-12-16).R,
%   R/oxygen.R
\name{optode.fixer}
\alias{optode.fixer}
\title{Optode ESM2 oxygen parsing correction}
\usage{
optode.fixer(O2SAT, O2CONC, salinity = 0)

optode.fixer(O2SAT, O2CONC, salinity = 0)
}
\arguments{
\item{O2SAT}{miss-parsed TEMP}

\item{O2CONC}{miss-parsed O2}

\item{depth}{depth of sensor in meters}

\item{optode_salinity}{internal optode setting default is 0}

\item{O2SAT}{miss-parsed TEMP}

\item{O2CONC}{miss-parsed O2}

\item{depth}{depth of sensor in meters}

\item{optode_salinity}{internal optode setting default is 0}
}
\value{
O2CONC oxygen concentration in mmol m-3, (umol / l) or ml/l

O2CONC oxygen concentration in mmol m-3, (umol / l) or ml/l
}
\description{
Optode ESM2 oxygen parsing correction

Optode ESM2 oxygen parsing correction
}
