% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{optode.lagcorrect_hahn}
\alias{optode.lagcorrect_hahn}
\title{Hahn optode lag correction}
\usage{
optode.lagcorrect_hahn(
  dateTime,
  TCphase,
  temp,
  coefs,
  tau = NA,
  tau_DO = c(14.8, -0.4),
  filter = T
)
}
\arguments{
\item{dateTime}{datetime vector (seconds or POSIXct)}

\item{TCphase}{optode phase}

\item{temp}{in-situ temperature}

\item{tau}{default = NA}

\item{tau_DO}{default is c(14.8, -0.4)}

\item{filter}{default = True to apply a median pre-filter}
}
\value{
lagged phase
}
\description{
Implements the Hahn optode lag correction, as used in UEA seaglider toolbox
}
