% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{optode.phaseCalc}
\alias{optode.phaseCalc}
\title{Optode phase}
\usage{
optode.phaseCalc(phase, Temp, coefs, silent = F)
}
\arguments{
\item{Temp}{vector of temperature (C)}

\item{coefs}{list (or data.frame) consisting of C0..C6 coefs, see examples}

\item{silent}{if True function won't print out the equation and batch numbers when called}

\item{DPhase}{vector of phase values}
}
\value{
oxygen concentration in mmol m-3
}
\description{
Uses supplied optode calibration coefficients to calculate oxygen concentration from Dphase/Calphase and temperature.
}
\details{
For mk1 optodes provide Dphase
for mk2 without SVU multipoint calibration provide Calphase (not TCphase!)
remember, for optodes > #1000 two point calibrations will only affect the linear transformation coefs.
}
\examples{

# for SVU multipoint calibrated optodes:
SVU_coef = list(batch="1517M", coef="SVU", C0=0.002757413, C1=0.000115132, C2=2.34E-06, C3=234.4436, C4=-0.3752192, C5=-45.60156, C6=4.630104)
optode.phaseCalc(45, 10, coefs=SVU_coef)

# For standard 4330, 4831, 4835 optodes:

# For standard 3835 optodes:
coefs = list(
  batch = 1707,
  coef = "mk1",
  C0 = c(5326.5, -192.117, 4.14357, -0.0378695),
  C1 = c(-292.068, 9.71993, -0.214295, 0.00200778),
  C2 = c(6.47595, -0.19808, 0.0044994, -0.0000431),
  C3 = c(-0.0669288, 0.00188066, -0.0000442, 0.000000428),
  C4 = c(0.000265042, -0.00000683, 0.000000167, -1.62E-09))
optode.phaseCalc(30, 10, coefs=coefs)
}
