% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{optode.tau}
\alias{optode.tau}
\title{Optode tau}
\usage{
optode.tau(temp, IL, type = c("fast", "standard", "SBE63"))
}
\arguments{
\item{temp}{in-situ temperature}

\item{IL}{boundary layer thickness}

\item{type}{either "fast", "standard" or "SBE63"}
}
\value{

}
\description{
Calculates Tau as per Bittig & Korsinger 2017
}
\details{
foil thickness (Im)
100 µm for the Aanderaa optodes with standard foil
50 µm for the Aanderaa 4330F optodes with fast response foil

3835, 4330 tau/s = 25 s
4330F tau/s = 8 s
RINKO tau/s = 0.4 s (gas phase)

Boundary layer thickness (IL) - dependent on platform
CTD-mounted optodes (range 20 – 50 µm), close to 40 at 1 dbar s-1
glider ~
210 - (110 / 0.095) * vel # if vel < 0.95 dbar s-1
40 + (60 / 0.905) * (1-vel) # if vel > 0.095 dbar s-1
}
