% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{oxygen.air_conc}
\alias{oxygen.air_conc}
\title{Oxygen in-air concentration}
\usage{
oxygen.air_conc(TEMP, AIRPRS, RH = NA, DTEMP = NA, return_conc = T)
}
\arguments{
\item{TEMP}{air temperature in Celsius}

\item{AIRPRS}{air pressure in mbar (hPa)}

\item{RH}{relative humidity in \\%}

\item{DTEMP}{dewpoint temperature in Celsius}

\item{return_conc}{if True (default) return concentration, otherwise returns partial pressure in hPa}
}
\value{
oxygen concentration in mmol m-3
}
\description{
Calculates in-air concentration of oxygen, given temperature, pressure and relative humidity.
If only given dewpoint will use Sargent (1980) approximation to derive relative humidity
}
\examples{
oxygen.air_conc(10, 1019.5, 80, return_con=F) # partial pressure of oxygen at 10oC and 1019.5 hPa local air pressure and 80\% relative humidity
}
\references{
Bittig, H.C., Körtzinger, A., Neill, C., van Ooijen, E., Plant, J.N., Hahn, J., Johnson, K.S., Yang, B., Emerson, S.R., 2018. Oxygen Optode Sensors: Principle, Characterization, Calibration, and Application in the Ocean. Frontiers in Marine Science 4, 1–25. https://doi.org/10/ggc8m9
}
\keyword{oxygen}
