% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{oxygen.sat}
\alias{oxygen.sat}
\title{Equilibrium Oxygen saturation concentration}
\usage{
oxygen.sat(temp, salinity, unit = "mmolm", P = 0, p_atm = 1013.25)
}
\arguments{
\item{temp}{numeric vector of water temperature in degrees Celsius}

\item{salinity}{numeric vector of salinity (PSU)}

\item{unit}{"mmolm" for mmol m-3 (default), "mgl" for mg l-1 or "umolkg" for umol kg-1.}

\item{P}{hydrosatic pressure in dbar (default = 0)}

\item{p_atm}{atmospheric (air) pressure in hPa (default = 1013.25)}
}
\value{
vector of saturation concentration in mmol m-3
}
\description{
Calculates oxygen saturation concentration in equilibrium with the atmosphere
as per Garcia & Gordon, 1992
}
\details{
conversions via SCOR WG 142
}
\examples{
oxygen.sat(10, 35)  # saturation concentration at 10 degrees and 35 salinity
oxygen.sat(10, 35, "molkg")  # saturation concentration at 10 degrees and 35 salinity in umol kg-1
}
\keyword{oxygen}
