% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{oxygen.winkler}
\alias{oxygen.winkler}
\title{Winkler titration equation}
\usage{
oxygen.winkler(Vthio, Vblank, Mthio, Vbot, Tfix, Tcal = 20, Vreag = 2)
}
\arguments{
\item{Vthio}{volume of thiosulfate at endpoint (ml)}

\item{Vblank}{V1 - V2 for blank (ml)}

\item{Mthio}{Molarity of thiosulfate (typically 0.1)}

\item{Vbot}{volume of sample bottle (ml)}

\item{Tfix}{temperature of sample when reagents added}

\item{Tcal}{temperature of thiosulfate during standardisation (default = 20)}

\item{Vreag}{volume of reagents added to sample (default = 2ml)}
}
\value{
oxygen concentration in \eqn{\mu*mol l^{-1}}
}
\description{
Calculates oxygen concentration from Winkler titation
}
\examples{
oxygen.winkler(1.2807, 0.001, 0.11801, 143.34, 11.5, 20, 2)

}
