% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{PAR_from_Voltage}
\alias{PAR_from_Voltage}
\title{Calculate Cefas LiCor PAR from voltage}
\usage{
PAR_from_Voltage(x, factor, offset)
}
\arguments{
\item{offset}{}
}
\description{
Calculates PAR from Licor Li-198 sensors
 fitted with Cefas inline low-light amplifier.
 Refer to calibration documents for factor and offset.
}

