% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeseriesChartsQA.R
\name{plot.sbts}
\alias{plot.sbts}
\title{SmartBuoy time series plots}
\usage{
\method{plot}{sbts}(deploymentGroup, parcode, yr = year(now()),
  parcode2 = FALSE, CT_temp_only = TRUE, return_data = FALSE,
  style = "dygraph", db_name = "smartbuoydblive")
}
\arguments{
\item{deploymentGroup}{character string matching SmartBuoy deployment group (case sensitive)}

\item{parcode}{character string matching parameter code of interested e.g. 'TEMP' or 'O2CONC'}

\item{yr}{integer vector corrisponding to year or years of interest e.g. c(2014, 2015), defaults to current year.}

\item{parcode2}{TODO}

\item{CT_temp_only}{optional boolean, if True and parcode = 'TEMP' only CT derived temperature data is returned}

\item{return_data}{optional boolean, if True timeseries data is returned rather than a plot}

\item{db_name}{optional character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
dygraph or ggplot object depending on style, or if return_data is True a data.table for the timeseries.
}
\description{
Fetches time series data for a deployment group
}
\details{
TODO
}
\examples{
plot.sbts('DOWSING', 'O2CONC')
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}

