% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{printvars}
\alias{printvars}
\title{print the variables in a netcdf (ncdump lite)}
\usage{
printvars(ncfile)
}
\arguments{
\item{ncfile}{}
}
\value{
variable names
}
\description{
print the variables in a netcdf (ncdump lite)
}
