% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.binning}
\alias{profiler.binning}
\title{ESM2 profiler depth binning}
\usage{
profiler.binning(x, bin_depth = 0.5, use_cast = "UP")
}
\arguments{
\item{x}{data.frame matching output from fetch.profiler}

\item{bin_depth}{numeric vector determining depth binning interval, default is 0.5}

\item{use_cast}{character string matching matching cast required, options are 'UP', 'DOWN' and 'BOTH'}
}
\value{
character vector of Cruise Id's
}
\description{
Binns profiler data into depth bins
}
\details{
TODO
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}

