% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.cruiselist}
\alias{profiler.cruiselist}
\title{ESM2 profiler cruise id queryer}
\usage{
profiler.cruiselist(yr = "ALL", db_name = "smartbuoydblive")
}
\arguments{
\item{yr}{integer specifing a year to limit the search, default is 'ALL'}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
character vector of Cruise Id's
}
\description{
Fetches lists of cruise id where ESM2 profiler data exists.
}
\details{
This function querys the Smartbuoy database and returns a list of valid cruise IDs matching the supplied critiera where ESM2
profiler data is available. The v_CtdProfile_AllData table is used, as such private data will not be available to this function.
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}

