% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.eu_depth}
\alias{profiler.eu_depth}
\title{estimate euphotic depth from PAR profile}
\usage{
profiler.eu_depth(depth, par, surf_par = NA)
}
\arguments{
\item{depth}{vector of depth values}

\item{par}{vector of par values}

\item{surf_par}{if not supplied attempts to find PAR at 0m depth}
}
\value{
euphotic depth (m)
}
\description{
estimate euphotic depth from PAR profile
}

