% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{profiler.fetch}
\alias{profiler.fetch}
\title{ESM2 profiler data query tool}
\usage{
profiler.fetch(
  cruiseID = NA,
  profiler = NA,
  after = NA,
  before = NA,
  area = NA,
  parameters = c("TEMP", "SAL", "FTU", "O2CONC", "PAR", "FLUORS"),
  RQ0 = TRUE,
  ct_temp_only = TRUE,
  min_QA_reached = TRUE,
  privateData = FALSE,
  debug = FALSE,
  db_name = "smartbuoydblive"
)
}
\arguments{
\item{cruiseID}{optional cruise ID string, e.g. "CEND_02_14", if provided only data from this cruise will be returned.}

\item{profiler}{optional profiler name string, e.g. "PR009", if provided only data from this profiler will be returned.}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{area}{optional area vector consisting of 4 elements, minimum Latitude, minimum Longitude, maximum Latitude, maximum Longitude e.g. c(52, -3.5, 53, -4)}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC', 'PAR', 'FLUORS')}

\item{RQ0}{boolean, if True only data where result quality = 0 is returned, i.e. good data. default is True}

\item{ct_temp_only}{boolean, if True all non-FSI temperature data is discarded, default is False.}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{privateData}{boolean, if True will return non-public data (e.g. commerical contracts), default is False.}

\item{debug}{boolean, if True will return raw query output}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.frame with returned data in "long" format or error string if no data returned
}
\description{
Fetches CTD data from ESM2 profiler records
}
\details{
This function querys the Smartbuoy database and returns ESM2 profiler data matching the provided critiera.
Note that startTime is the time at which the instrument started logging.
remember to do \code{options(digits.secs=3)} if you want to display POSIXct miliseconds.

If using RQ0 = F QA flags are returned:
\itemize{
\item 0 - Data passed all active QA checks
\item 1 - spike threshold exceeded
\item 2 - below minimum for parameter
\item 3 - above maximum for parameter
\item 4 - biofouled sensor
\item 12 - sensor malfunction
\item 13 - Tpeak anomaly
\item 14 - dependency failure
\item 15 - WMS bag not spiked with mercuric chloride
\item 100 - sensor is out of water - 1
\item 200 - results outside limits (level 2) - 1
\item 16 - low battery voltage
\item 17 - sensor not connected
\item 18 - deviation from rolling mean
\item 19 - deviation from neighbouring point
\item 201 - results outside specified concentration range (level 2) - 1
\item 20 - sensor flooded
\item 21 - possible chlorophyll fluorescence quenching
\item 255 - Backup Results
\item 22 - Low volume sample
\item 5 - removed during Median filtering
\item 6 - light intrusion effect on optical sensor
\item 7 - weed or debris on sensor / sensor obscured
\item 8 - hunting error on auto-ranging sensor
\item 9 - suspicious data (unknown cause)
\item 10 - discrepancy between two data sets
\item 11 - sensor platform out of position
}
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}
