% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.kd}
\alias{profiler.kd}
\title{Estimate Kd from profile}
\usage{
profiler.kd(depth, par)
}
\arguments{
\item{depth}{vector of depth values}

\item{par}{vector of par values}
}
\value{
kd (m-1)
}
\description{
Estimates Kd PAR from downwelling PAR and depth using log par method.
}
