% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{profiler.match_ferrybox}
\alias{profiler.match_ferrybox}
\title{ESM2 profiler ferrybox matching}
\usage{
profiler.match_ferrybox(
  cruiseID = NA,
  parameters = c("TEMP"),
  min_QA_reached = F,
  ferrybox_db_name = "ferrybox",
  smartbuoy_db_name = "smartbuoydblive"
)
}
\arguments{
\item{cruiseID}{cruise ID string, e.g. "CEND_02_14"}

\item{parameters}{vector of parameter code names, defaults to c('TEMP')}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{ferrybox_db_name}{character string matching ODBC data source name, defaults to 'ferrybox'}

\item{smartbuoy_db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
list containing matched data and regression ggplot
}
\description{
Fetches and matches profiler and ferrybox data
}
\details{
TODO
}
\keyword{QA}
\keyword{ctd}
\keyword{ferrybox}
\keyword{profiler}
