% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{profiler.odz}
\alias{profiler.odz}
\title{Estimate optical depths from PAR profile}
\usage{
profiler.odz(depth, par)
}
\arguments{
\item{depth}{vector of depth values}

\item{par}{vector of par values}
}
\value{
vector of optical depth zone for each depth
}
\description{
Estimate optical depths from PAR profile
}
