% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{profiler.QA3app}
\alias{profiler.QA3app}
\title{Profiler QA3}
\usage{
profiler.QA3app()
}
\value{
starts the profiler QA3 app
}
\description{
Starts the shiny profiler QA3 app.
This application helps generate QA3 regressions using niskin samples from profiles.
Have niskin sample data to hand!.
}
