% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.salinityQA}
\alias{profiler.salinityQA}
\title{profiler salinity QA}
\usage{
profiler.salinityQA(dateTime, depth, cruise, bottlesal, bin_height = 1,
  method = round, mode = "time_matched", station = NA,
  time_tolerance = 120, min_count = 8, use_cast = "UP")
}
\arguments{
\item{dateTime}{}

\item{depth}{}

\item{cruise}{}

\item{bottlesal}{}

\item{bin_height}{}

\item{method}{}

\item{mode}{}

\item{station}{}

\item{time_tolerance}{}

\item{min_count}{}

\item{use_cast}{}
}
\value{
list containing data and a plot
}
\description{
profiler salinity QA
}

