% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.split}
\alias{profiler.split}
\title{Split ESM2 profiler data in distinct profiles}
\usage{
profiler.split(
  x,
  t_threshold = 180,
  z_threshold = 0.25,
  use_prs = F,
  airprs = 10
)
}
\arguments{
\item{x}{profiler data.table, such as that generated by \code{fetch.profiler}}

\item{t_threshold}{time interval in seconds, default is 180 s}

\item{z_threshold}{depth / pressure threshold in either meters or dbar}

\item{use_prs}{if true then function will use PRSADC and the \code{airprs} parameter to determine depth, otherwise uses depth column which needs to already be calculated}

\item{airprs}{in-air pressure reading, if using \code{use_prs} typically around 10 dbar}
}
\value{
data.table containing profiler data with startTimes updated and "profile" column added
}
\description{
Split profiles for ESM2 profiler data.
If an ESM2 as been left sampling between station, or a station as been repeated the multiple casts will have the same "startTime"
}
\details{
This function assumes that the ESM2 leaves the water between stations, uses that and a minimum time difference to mark seperate stations.
Specify a threshold for time-between-stations (\code{t_threshold}), and a threshold for when it is out of the water (\code{z_threshold})

A pure "out of water" threshold won't work because an ESM2 often bobs when near the surface
}
