% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prop_dt}
\alias{prop_dt}
\title{propagate errors for data.table}
\usage{
prop_dt(expr, dat, mode = c("prop", "MC"), n = 1e+06)
}
\arguments{
\item{expr}{an expression, such as \code{expression(x / y)}}

\item{dat}{a matrix denoting the mean and standard deviation, see examples}

\item{mode}{either \code{"prop"} or \verb{"MC}}

\item{n}{number of iterations for Monte-Carlo, minimum is 10000}
}
\value{
a list of mean and standard deviations
}
\description{
This function serves as a wrapper for \code{propagate::propagate} to make it easier to calculate errors in a data.table
}
\details{
Errors are calculated though a 2nd order Taylor expansion (\code{mode = "prop"}) or though Monte-Carlo (\code{mode = "MC"}).
}
\examples{
dx = data.table(
  index = c(500, 1000),
  A = c(0, 5),
  A_sd = c(0.1, 0.1),
  B = c(1, 9),
  B_sd = c(0.1, 0.1))

eq = expression(A + B^2)

dx[, c("mean", "sd") := prop_dt(eq, cbind("A" = c(A, A_sd), "B" = c(B, B_sd))), by=index]

# Can also be applied to a standard function, if you create a wrapper
d2 = data.table(
  time = c(1, 2),
  temp = c(20, 5),
  temp_sd = c(0.2, 0.1),
  sal = c(35.5, 34.5),
  sal_sd = c(0.01, 0.05))

f = function(temp, sal){
  # a wrapper is needed if you don't supply every variable, and you can't have non-numeric variables
  oxygen.sat(temp, sal)
}

d2[, c("mean", "sd") := prop_dt(f, cbind("temp" = c(temp, temp_sd), "sal" = c(sal, sal_sd))), by=time]
}
