% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qpercentunif}
\alias{qpercentunif}
\title{quantile uniform random numbers from percentage}
\usage{
qpercentunif(p, mean, percent)
}
\arguments{
\item{p}{a vector of probabilies}

\item{mean}{mean of observations}

\item{percent}{percentage of mean}
}
\value{
numeric vector of random deviates
}
\description{
This is a wrapper for qunif that calculates the upper and lower bounds based on a percentage of a given mean
}
\examples{
rpercentunif(10, 50, 10)
}
