% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rate_of_change}
\alias{rate_of_change}
\title{Calculate rate of change}
\usage{
rate_of_change(dateTime, value)
}
\arguments{
\item{dateTime}{POSIXct vector}

\item{value}{numeric vector}
}
\value{
vector of rate of change in unit/second
}
\description{
Calculates rate of change given datetime and value
}
