% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.BODC_ctd_ascii}
\alias{read.BODC_ctd_ascii}
\title{Read BODC CTD ascii (lst) files}
\usage{
x = lapply(list.files(folder, full.names = T, pattern = "*.lst"), read.BODC_ctd_asci)
d = rbindlist(x, fill = T)
}
\arguments{
\item{file}{string indicating BODC file location}

\item{stripAgg}{if True (default) aggregation is used}
}
\value{
data.frame (data.table)
}
\description{
Reads ascii CTD files from BODC, attempts to extract metadata
}
