% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.CTDQC}
\alias{read.CTDQC}
\title{Read CTDQC data}
\usage{
read.CTDQC(session, type = c("data", "untrimmed"))
}
\arguments{
\item{session}{i.e. that loaded with load("CTDQC.rdata")}

\item{type}{either "data" for the trimmed downcasts or "untrimmed" for everything}
}
\value{
data.table of all CTDs
}
\description{
Extracts in a nice data.table format the nested oce CTD casts from a CTDQC session
}
