% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{read.ecmwf}
\alias{read.ecmwf}
\title{Read and convert ECMWF netcdf to data.table}
\usage{
read.ecmwf(file, convert_units = F)
}
\arguments{
\item{file}{name of ECMWF netcdf file}

\item{convert_units}{if True convert to more typical units (i.e. Celcius / mbar), default is False.}
}
\value{
data.table containing all netcdf variables in wide format
}
\description{
Read and convert ECMWF netcdf to data.table
}
