% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.ego}
\alias{read.ego}
\title{Read BODC ego glider files (e.g. slocum)}
\usage{
read.ego(ncfile, vars = c("DOXY", "PRES", "TEMP", "CNDC", "CHLA", "BBP700"))
}
\arguments{
\item{ncfile}{ego netcdf file}

\item{vars}{character vector of variables to extract, e.g. c("DOXY", "PRES")}
}
\value{
data.table containing extracted fields
}
\description{
Read BODC ego glider files (e.g. slocum)
}
