% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ESMxburst}
\alias{read.ESMxburst}
\title{Read ESMx Burst files}
\usage{
read.ESMxburst(file)
}
\arguments{
\item{file}{}
}
\description{
Read ESMx Burst files
}
\examples{
flist = list.files(pattern = "Burst*") # get the files e.g. "1001_Burst 2.csv"
d = lapply(flist, read.ESMxburst)
d = rbindlist(d)[order(dateTime)]
}
