% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.10min}
\alias{read.ferrybox.10min}
\title{Ferrybox 10min file data reader}
\usage{
read.ferrybox.10min(folder)
}
\arguments{
\item{folder}{folder containing 10min files}
}
\value{
data.frame (data.table) of processed 10min files
}
\description{
reads JENA-4H 10min files
}
\details{
TODO
}
\keyword{10minfile}
\keyword{ferrybox}

