% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.conlog}
\alias{read.ferrybox.conlog}
\title{Ferrybox conlog data reader}
\usage{
read.ferrybox.conlog(conlog)
}
\arguments{
\item{conlog}{character string indicating device data file}
}
\value{
data.frame
}
\description{
reads DAQ server conlog data files
}
\details{
TODO
requires intensive string operations on many rows and thus fairly slow
}
\keyword{conlog}
\keyword{ferrybox}
