% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.devdata}
\alias{read.ferrybox.devdata}
\title{Ferrybox device data reader}
\usage{
read.ferrybox.devdata(devdata_folder, pivot = F, print = T)
}
\arguments{
\item{devdata_folder}{character string indicating path to folder containing device data files}

\item{pivot}{optional boolean indicating if returned table should be recast into 'wide' format}

\item{print}{optional boolean indicating if current file should be printed, default is True.}
}
\value{
data.table
}
\description{
reads ferrybox device data files
}
\details{
Ferrybox quality codes:
\itemize{
\item 0 = good,
\item 1 = no data,
\item 2 = over value,
\item 4 = under value,
\item 8 = sensor timeout
\item 64 = clean cycle,
\item 128 = standby,
\item 256 = empty,
\item 512 = error,
\item 1024 = undefined,
\item 2048 = simulation mode
}

sometimes \0 characters appear in data, to sanitise run sed from a bash terminal
}
\keyword{ferrybox}
