% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.liveAquire}
\alias{read.liveAquire}
\title{Read (and check) ESM2 LiveAquire}
\usage{
read.liveAquire(f, plot = T)
}
\arguments{
\item{f}{filename}

\item{plot}{if True (default) plots the liveaquire data}
}
\value{
ggplot or data.table depending on value of "plot"
}
\description{
Function to parse and optionally plot liveaquire data
}
