% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read.NEMA_GGA}
\alias{read.NEMA_GGA}
\title{Parse a NMEA GGA GPS file}
\usage{
read.NEMA_GGA(file, date_origin = "2000-01-01")
}
\arguments{
\item{date_origin}{default "2000-01-01"}
}
\value{
data.table with processed positions
}
\description{
Tool to parse a text log containing GGA NMEA strings
Note that GGA has a timestamp but not a date, so the date origin needs to be supplied
}
\details{
Tool returns gpsTime (UTC), lat and lon in decimal degrees, number of satelites.
HDOP and altitude in meters.
}
\examples{
 gps = read.NEMA_GGA("teraterm.log", date_origin = "2021-10-02")
}
