% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R, R/misc.R
\name{read.NMEA_GGA}
\alias{read.NMEA_GGA}
\title{read and parse NMEA GGA log file}
\usage{
read.NMEA_GGA(file, date_origin = "2000-01-01")

read.NMEA_GGA(file, date_origin = "2000-01-01")
}
\arguments{
\item{file}{filename}

\item{date_origin}{default "2000-01-01"}

\item{only_good}{if True (default) discard lines with bad GPS fixes}
}
\value{
data.table containing, gpsTime, lat, lon, number of satelites, HDOP (meters) and altitude (meters)

data.table with processed positions
}
\description{
Tool will parse a NMEA log file, and decode each line with a "$xxGGA" string
Note that GGA does not contain the date, only UTC time, so a date origin must be supplied for each file.
If your log file spans multiple days, make sure you handle the roll-over at midnight!

Tool to parse a text log containing GGA NMEA strings
Note that GGA has a timestamp but not a date, so the date origin needs to be supplied
}
\details{
This tool also ignores any prefix to the $xxGGA string start, so preceding Teraterm timestamps are ignored.

Tool returns gpsTime (UTC), lat and lon in decimal degrees, number of satelites.
HDOP and altitude in meters.
}
\examples{
read.NMEA_GGA("teraterm.log", "2020-01-01")
 gps = read.NMEA_GGA("teraterm.log", date_origin = "2021-10-02")
}
