% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.nutrients}
\alias{read.nutrients}
\title{Read LIMS style nutrients excel spreadsheet}
\usage{
read.nutrients(filename)
}
\arguments{
\item{filename}{single xlsx filename}
}
\value{
long format data.table
}
\description{
Tool to parse the "Samples" sheet from a modern nutrients lab spreadsheet (.xlsx only)
}
