% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.seaglider_eng}
\alias{read.seaglider_eng}
\title{Read seaglider .eng files}
\usage{
read.seaglider_eng(folder = NA, files = NA)
}
\arguments{
\item{folder}{folder containing .eng}

\item{files}{single files if not using folder}
}
\value{
data.table containing extracted fields
}
\description{
Read seaglider .eng files
}
