% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.seaglider_log}
\alias{read.seaglider_log}
\title{read seaglider logfiles, extracts GPD, altimetry, battery data, and targets}
\usage{
read.seaglider_log(glider_folder)
}
\arguments{
\item{glider_folder}{}
}
\value{
data.table containing processed logfiles
}
\description{
read seaglider logfiles, extracts GPD, altimetry, battery data, and targets
}
