% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.seaglider_toolbox_ts_nc}
\alias{read.seaglider_toolbox_ts_nc}
\title{read UEA seaglider toolbox timeseries netcdf}
\usage{
read.seaglider_toolbox_ts_nc(
  ncfile,
  variables = c("sigma0", "salinity", "oxygen", "temp")
)
}
\arguments{
\item{ncfile}{UEA glider netcdf timeseries file}

\item{variables}{varibles to extract as named in netcdf, e.g. c("salinity", "temp")}
}
\value{
data.table containing extracted fields
}
\description{
extracts data from UEA seaglider toolbox timeseries netcdf output
always includes dive, direction, pressure, lat and lon
}
