% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.SmartBuoyLiveAquireExport}
\alias{read.SmartBuoyLiveAquireExport}
\title{SmartBuoy live aquire export reader}
\usage{
read.SmartBuoyLiveAquireExport(
  file = svDialogs::dlgOpen(title = "Open LiveAquire file...")$res
)
}
\arguments{
\item{file}{location of live aquire file, leave blank to use dialog box}
}
\value{
a list containing a data frame and ggplot
}
\description{
reads live aquire files exported from smartbuoydb and plots output
}
\details{
This function reads text files generated by the SmartBuoy interface live aquire.
}
\keyword{esm2}
\keyword{smartbuoy}
