% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.teraterm_log}
\alias{read.teraterm_log}
\title{Read teraterm datetime log}
\usage{
read.teraterm_log(f)
}
\arguments{
\item{f}{filename}
}
\value{
data.table
}
\description{
Teraterm has a timestamp option, however it's awkward to parse, this function takes a log file created by teraterm,
reads it as a table and extracts the dateTime column.
}
\details{
You will need to clean the file to remove commands or badly formatted rows.
}
