% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ULP000}
\alias{read.ULP000}
\title{Process ESM2 .000 files}
\usage{
read.ULP000(file)
}
\arguments{
\item{file}{an extracted ESM2 .000 filename

to unpack a .ULP use the \code{unulp.exe} with the following command: \code{ulunp -i *.ULP}.}
}
\value{
data.table
}
\description{
Process ESM2 .000 files
}
