% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavenet.R
\name{read.wavenet_sbd_info}
\alias{read.wavenet_sbd_info}
\title{Decode DWR3 sbd file info}
\usage{
read.wavenet_sbd_info(sbd, debug = F)
}
\arguments{
\item{sbd}{waverider sbd filename}

\item{debug}{if True, print matching messages}
}
\value{
list with named elements of lat, lon and battery (in weeks)
}
\description{
Decode DWR3 sbd file info
}
