% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{RH_from_dewtemp}
\alias{RH_from_dewtemp}
\title{Calculate relative humidity from air temperature and dewpoint}
\usage{
RH_from_dewtemp(t2m, d2m, unit = "C")
}
\arguments{
\item{t2m}{air temperature}

\item{d2m}{dew point temperature}

\item{unit}{"K" or "C" (default)}
}
\value{
relative humidity (\%)
}
\description{
Calculate relative humidity from air temperature and dewpoint
}
