% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{rinko.o2}
\alias{rinko.o2}
\title{Calculate RINKO oxygen from voltage}
\usage{
rinko.o2(
  V,
  t,
  S,
  oC = list(A = -42.34162, B = +127.6475, C = -0.3677435, D = +0.01137, E = +0.0046, F =
    +7.57e-05),
  p = 10.1325,
  G = 0,
  H = 1
)
}
\arguments{
\item{V}{output oxygen voltage}

\item{t}{temperature from rinko_temp}

\item{oC}{list of named calibration coefs A-F, defaults to #0263 ARO-CAV}

\item{p}{in-situ pressure in dBar, default to atmospheric (10.1325)}

\item{G}{alpha calibration coef}

\item{H}{beta calibration coef}
}
\value{
vector of RINKO oxygen in mmol m-3
}
\description{
Calculate RINKO oxygen from voltage
}
