% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{rinko.p}
\alias{rinko.p}
\title{Calculate RINKO saturation from voltage}
\usage{
rinko.p(
  V,
  t,
  S,
  oC = list(A = -42.34162, B = +127.6475, C = -0.3677435, D = +0.01137, E = +0.0046, F
    = +7.57e-05, G = +0, H = +1),
  p = 0,
  G = 0,
  H = 1
)
}
\arguments{
\item{V}{output oxygen voltage}

\item{t}{temperature from rinko_temp}

\item{oC}{list of named calibration coefs A-F, defaults to #0263 ARO-CAV}

\item{p}{hydrostatic pressure in dBar, default = 0}

\item{G}{alpha calibration coef}

\item{H}{beta calibration coef}
}
\value{
vector of RINKO oxygen in \\% saturation
}
\description{
Calculate RINKO saturation from voltage
}
