% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{rinko.temp}
\alias{rinko.temp}
\title{Calculate RINKO temperature from voltage}
\usage{
rinko.temp(
  V,
  tC = list(A = -5.326887, B = +16.63288, C = -2.123968, D = +0.4543014)
)
}
\arguments{
\item{V}{measured RINKO output in volts}

\item{tC}{list of named coef, defaults to those of #0263 ARO-CAV}
}
\value{
vector of temperatures
}
\description{
Calculate RINKO temperature from voltage
}
