% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{round_minute}
\alias{round_minute}
\title{Round dateTime to x minutes}
\usage{
round_minute(x, minutes = 30)
}
\arguments{
\item{x}{vector of POSIXct datetimes}

\item{minutes}{integer minutes to round to, default is 30}
}
\value{
vector of rounded POSIXct
}
\description{
Round dateTime to x minutes
}
