% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rpercentunif}
\alias{rpercentunif}
\title{uniform random numbers from percentage}
\usage{
rpercentunif(n, mean, percent)
}
\arguments{
\item{n}{number of observations}

\item{mean}{mean of observations}

\item{percent}{percentage of mean}
}
\value{
numeric vector of random deviates
}
\description{
This is a wrapper for runif that calculates the upper and lower bounds based on a percentage of a given mean
}
\examples{
rpercentunif(10, 50, 10)
}
