% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{SAL_from_CT}
\alias{SAL_from_CT}
\title{Calculate salinity}
\usage{
SAL_from_CT(Cond, t, p = max(0, P - 1.013253), P = 1.013253)
}
\arguments{
\item{Cond}{Conductivity in Sm, (SmartBuoy = mmoh/cm = 0.1 Sm)}

\item{t}{temperature}

\item{p}{gauge pressure in bar (reference to ambient)}

\item{P}{true pressure in bar}
}
\value{
salinity
}
\description{
calculate salinity from conductivity, temperature and depth.
Adapted from marelac package by Karline Soetaert using
Fofonoff NP and Millard RC Jr, 1983. Algorithms for computation of fundamental properties of
seawater. UNESCO technical papers in marine science, 44, 53 pp.
http://unesdoc.unesco.org/images/0005/000598/059832EB.pdf
cran.r-project.org/web/packages/marelac
}
\keyword{CTD}
\keyword{conductivity}
\keyword{salinity}
