% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{saturation_vapour_pressure}
\alias{saturation_vapour_pressure}
\title{Calculate water vapour saturation pressure}
\usage{
saturation_vapour_pressure(TEMP, unit = "C", method = "Wagner")
}
\arguments{
\item{TEMP}{vector of temperature}

\item{unit}{"C" default or "K"}

\item{method}{options are "Wagner", "Buck" and "Weiss"}
}
\value{
saturation vapour pressure in hPa (mBar)
}
\description{
Calculate water vapour saturation pressure
}
