% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeseriesChartsQA.R
\name{sbts}
\alias{sbts}
\title{SmartBuoy time series plots}
\usage{
sbts(deploymentGroup, parcode, yr = year(now()), CT_temp_only = TRUE,
  style = "dygraph", include_telemetry = TRUE, day_flu_only = TRUE,
  db_name = "smartbuoydblive")
}
\arguments{
\item{deploymentGroup}{character string matching SmartBuoy deployment group (case sensitive)}

\item{parcode}{character string matching parameter code of interested e.g. 'TEMP' or 'O2CONC'}

\item{yr}{integer vector corrisponding to year or years of interest e.g. c(2014, 2015), defaults to current year.}

\item{CT_temp_only}{optional boolean, if True and parcode = 'TEMP' only CT derived temperature data is returned}

\item{include_telemetry}{optional boolean, if True (default) telemetry data is also used for the timeseries}

\item{db_name}{optional character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
dygraph or ggplot object depending on style, or if return_data is True a data.table for the timeseries.
}
\description{
Fetches time series data for a deployment group
}
\details{
TODO
}
\examples{
plot.sbts('DOWSING', 'O2CONC')
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}

