% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{seaglider.pressure}
\alias{seaglider.pressure}
\title{Calculate seaglider pressure}
\usage{
seaglider.pressure(depth)
}
\arguments{
\item{depth}{vector of seaglider depth (cm)}
}
\value{
vector of calibrated pressure in dbar
}
\description{
calculates true pressure from .eng file "depth" variable
}
