% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{seaglider.temp}
\alias{seaglider.temp}
\title{Calculate seaglider temperature}
\usage{
seaglider.temp(tempFreq, calib_file)
}
\arguments{
\item{tempFreq}{vector of seaglider SBE temperature frequency}

\item{calib_file}{path to seaglider calibration.mat file}
}
\value{
vector of calibrated temperature
}
\description{
Calculate seaglider temperature
}
