% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sec_axis_rescale}
\alias{sec_axis_rescale}
\title{rescale axis for use with ggplot2 sec_axis}
\usage{
sec_axis_rescale(x, y)
}
\arguments{
\item{x}{vector of variable to be scaled}

\item{y}{vector of variable to scale against}
}
\value{
rescaled vector
}
\description{
rescales variables to use with ggplot2 when using secondary axis,
perfect for a CTD profile.
}
\examples{
library(oce)
data("ctd") # example CTD profile from oce

x = as.data.frame(ctd@data)

ggplot(x) +
 geom_path(aes(salinity, pressure, color = "Salinity")) +
 geom_path(aes(sec_axis_rescale(temperature, salinity), pressure, color = "Temperature")) +
 scale_x_continuous("Salinity", sec.axis = sec_axis(bquote(Temperature~(degree*C)),
                                                    trans = ~ sec_axis_rescale(., x$temperature))) +
 scale_y_reverse("Pressure (dbar)")
}
