% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.fetch}
\alias{smartbuoy.fetch}
\title{SmartBuoy data query tool}
\usage{
smartbuoy.fetch(
  deployment = NA,
  deployment_group = NA,
  after = NA,
  before = NA,
  parameters = c("TEMP", "SAL", "FTU", "O2CONC", "FLUORS"),
  min_QA_reached = TRUE,
  RQ0 = TRUE,
  ct_temp_only = TRUE,
  averaging_period = NA,
  night_flu_only = FALSE,
  db_name = "smartbuoydblive"
)
}
\arguments{
\item{deployment}{optional string matching smartbuoy deployment}

\item{deployment_group}{optional string matching smartbuoy deployment_group}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC', 'FLUORS')}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{RQ0}{boolean, if True only data where result quality = 0 is returned, i.e. good data. default is True}

\item{ct_temp_only}{boolean, if True all non-FSI temperature data is discarded, default is False.}

\item{averaging_period, }{integer matching length in hours of averaging interval, i.e. 24 for daily means}

\item{night_flu_only}{optional boolean, if True (default) only night flu data will be returned}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table with returned data in "long" format or error string if no data returned
}
\description{
Fetches SmartBuoy data
}
\details{
This function queries the SmartBuoy database and returns a data.table

The averaging_period parameter does a simple time-stamp based block-mean.
i.e. it will include data from multiple sensors or buoys if they overlap.

If using RQ0 = F QA flags are returned:
\itemize{
\item 0 - Data passed all active QA checks
\item 1 - spike threshold exceeded
\item 2 - below minimum for parameter
\item 3 - above maximum for parameter
\item 4 - biofouled sensor
\item 12 - sensor malfunction
\item 13 - Tpeak anomaly
\item 14 - dependency failure
\item 15 - WMS bag not spiked with mercuric chloride
\item 100 - sensor is out of water - 1
\item 200 - results outside limits (level 2) - 1
\item 16 - low battery voltage
\item 17 - sensor not connected
\item 18 - deviation from rolling mean
\item 19 - deviation from neighbouring point
\item 201 - results outside specified concentration range (level 2) - 1
\item 20 - sensor flooded
\item 21 - possible chlorophyll fluorescence quenching
\item 255 - Backup Results
\item 22 - Low volume sample
\item 5 - removed during Median filtering
\item 6 - light intrusion effect on optical sensor
\item 7 - weed or debris on sensor / sensor obscured
\item 8 - hunting error on auto-ranging sensor
\item 9 - suspicious data (unknown cause)
\item 10 - discrepancy between two data sets
\item 11 - sensor platform out of position
}
}
\keyword{esm2}
\keyword{query}
\keyword{smartbuoy}
