% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.fetch_burst}
\alias{smartbuoy.fetch_burst}
\title{SmartBuoy burst level data query tool}
\usage{
smartbuoy.fetch_burst(deployment = NA, parameters = NA,
  db_name = "smartbuoydblive")
}
\arguments{
\item{deployment}{optional string matching smartbuoy deployment}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC', 'FLUORS')}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table with returned data in "long" format or error string if no data returned
}
\description{
Fetches burst level SmartBuoy data
}
\details{
This function querys the Smartbuoy database and returns
}
\keyword{esm2}
\keyword{query}
\keyword{smartbuoy}

