% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.parcodes}
\alias{smartbuoy.parcodes}
\title{SmartBuoy parameter codes}
\usage{
smartbuoy.parcodes(db_name = "smartbuoydblive")
}
\arguments{
\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table of parcodes
}
\description{
Fetches parameter codes from SmartBuoy database
}
\keyword{SmartBuoy}
