% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{smartbuoy.positions}
\alias{smartbuoy.positions}
\title{Deployment Map}
\usage{
smartbuoy.positions(db_name = "smartbuoydblive")
}
\arguments{
\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table of positions
}
\description{
Creates a map of MOS deployment sites
}
\details{
1 for SmartBuoy, 4 = Lander, 8 = Waverider
}
\keyword{SmartBuoy}

