% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.sensorCals}
\alias{smartbuoy.sensorCals}
\title{Sensor calibration query}
\usage{
smartbuoy.sensorCals(deployment = NA, deployment_group = NA,
  parameters = NA, db_name = "smartbuoydblive")
}
\arguments{
\item{deployment}{optional string vector denoting deployment names e.g. "DOWSING/042"}

\item{deployment_group}{optional string vector denoting deployment group names e.g. "DOWSING"}

\item{parameters}{optional parameter codes to match sensors by e.g. c("FLUORS", "FTU")}

\item{db_name}{optional character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.table containing query results
}
\description{
Tool for fetching serial numbers and calibration dates for SmartBuoy sensor
}
