% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{smartbuoy.telid}
\alias{smartbuoy.telid}
\title{Telemetry id lookup function}
\usage{
smartbuoy.telid(x)
}
\arguments{
\item{x}{a vector of telemetry ids}
}
\value{
a vector of matching sensor parameter names
}
\description{
Telemetry id lookup function
}

