% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeseriesChartsQA.R
\name{smartbuoy.TS}
\alias{smartbuoy.TS}
\title{SmartBuoy T/S plots}
\usage{
smartbuoy.TS(deployment, db_name = "smartbuoydblive")
}
\arguments{
\item{db_name}{optional character string matching ODBC data source name, defaults to 'smartbuoydblive'}

\item{deploymentcharacter}{string matching SmartBuoy deployment (case sensitive)}
}
\value{
ggplot object
}
\description{
Draws a T/S plot for a SmartBuoy Deployment
}
\details{
TODO
}
\keyword{esm2}

