% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{telids}
\alias{telids}
\title{telids}
\format{
A data frame with 96 rows and 3 variables:
\describe{
\item{\code{telid}}{The Id as represented on ULEEF and the database}
\item{\code{sensor_parameter}}{parameter code}
\item{\code{in_fb_loop}}{Is this telid used on the Cefas Endeavour ferrybox}
}
}
\usage{
telids
}
\description{
A table of the telemetry id's from the Cefas SmartBuoy database
}
\keyword{datasets}
