% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_glider}
\alias{theme_glider}
\title{Glider ggplot2 theme}
\usage{
theme_glider(
  base_size = 11,
  font_size = 12,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14,
  legend.position = "bottom"
)
}
\value{
ggplot2 theme
}
\description{
based on theme_bw, moves legend to the bottom, makes it wide, useful for matlab style glider plots
}
