% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidyMCMCts}
\alias{tidyMCMCts}
\title{Tidier for Stan timeseries}
\usage{
tidyMCMCts(fit)
}
\arguments{
\item{fit}{a stanfit object}
}
\value{
tidied data.table
}
\description{
Takes a stanfit object, calculates median, and 95 \% quantiles, includes Rhat and number of effective samples and includes a column for the observation number,
very useful for time-series data with large parameter vectors.
}
