% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{time_group}
\alias{time_group}
\title{Time group}
\usage{
time_group(x, fuzz = 0)
}
\arguments{
\item{x}{vector or timestamps}

\item{fuzz}{amount of extra leeway in seconds to give the function to avoid small gaps, default is 0}
}
\value{
vector of group indexes
}
\description{
given a timeseries with gaps, this returns a vector which can serve as a label for each unbroken period
}
\details{
in short, it calculates the median time between all observations,
and assigns a new group number when there is a gap.
}
