% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tstrgroup}
\alias{tstrgroup}
\title{split string by group}
\usage{
tstrgroup(x, pattern)
}
\arguments{
\item{x}{string}

\item{pattern}{a regex string with arbitrary number of matching groups}
}
\value{
a list containing the whole matched string, and each group
}
\description{
When you want to split a column in a data.table and extract various elements into new columns
}
\examples{
d = data.table(cruise = c("CEND 1/13", "HORK", "bon"), rep = 1:3)
d[, c("original", "v1", "v2", "v3") := tstrgroup(cruise, "(CEND) ([\\\\d\\\\w]+)/(\\\\d+)")] # 3 groups
}
