% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourmaps.R
\name{turbo}
\alias{turbo}
\title{Google "Turbo" colour map}
\usage{
turbo(n, start = 0, end = 1)
}
\arguments{
\item{n}{number of colours}

\item{start}{default = 0}

\item{end}{default = 1}
}
\description{
A colour map for those who really like Jet but know they can do better:
}
\details{
https://ai.googleblog.com/2019/08/turbo-improved-rainbow-colormap-for.html
}
\examples{

ggplot(faithfuld, aes(waiting, eruptions)) +
geom_raster(aes(fill = density)) +
scale_fill_gradientn(colors=turbo(256))
}
