% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{UKF}
\alias{UKF}
\title{Unscented Kalman filter (UKF)}
\usage{
UKF(
  y,
  mod,
  GGfunction,
  FFfunction,
  kappa = 0,
  sqrtMethod = "Cholesky",
  logLik = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{y}{data. The data can be a vector (one-dimensional measurements / # univariate time series)
or a matrix (multidimensional measurements /
multivariate time series). In the matrix case, each of the columns contains
the measurements on a single dimension (or, similarly, a single \link{univariate}
time series).
Missing values (NA's) are allowed.}

\item{mod}{a list with the components m0 (initial state estimates),}

\item{GGfunction}{a function with arguments x and k. The GGfunction specifies the
state transition function. The transition function is usually denoted as
f(x\link{k}, u\link{k}), where x are the states estimates and u the control input at
time step k.
See details below for how to handle control input in the GGfunction.}

\item{FFfunction}{a function with arguments x and k. The FFfunction specifies the
observation/measurement function. The measurement function is usually denoted as h(x\link{k}).}

\item{kappa}{a tuning parameter.}

\item{sqrtMethod}{specifies the method that will be used for computing the matrix,
square root. The method is either "Cholesky" (=Cholesky decomposition, the default) or svd" (=singular value decomposition).}

\item{logLik}{if TRUE, then return the negative loglikelihood of the specified model.}

\item{simplify}{if TRUE, then do not include the data (=y) in the output.}
}
\value{
list containing,
m = the filtered state estimates
C = the filtered state covariance
}
\description{
This function implements the Unscented Kalman Filter (UKF) as described
by Julier and Uhlmann in their 1997 paper "A New Extension of the Kalman Filter to Nonlinear Systems".
}
\details{
Note that this UKF assumes (time invariant) additive noise.
}
\references{
Julier & Uhlmann, 1997, A New Extension of the Kalman Filter to Nonlinear Systems
}
\author{
Stefan Gelissen
}
