% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{UKFforecast}
\alias{UKFforecast}
\title{UKF forecast}
\usage{
UKFforecast(filterData, nAhead = 1)
}
\arguments{
\item{filterData}{an object as returned by UKF.}

\item{nAhead}{the number of steps ahead for which to predict system states}
}
\description{
predicts future system states and observations for the Unscented Kalman Filter (UKF).
}
\author{
Stefan Gelissen
}
