% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKF.R
\name{UKFsmooth}
\alias{UKFsmooth}
\title{RTS smoother for the Unscented Kalman filter (UKF)}
\usage{
UKFsmooth(filterData)
}
\arguments{
\item{filterData}{output from UKF}
}
\value{
list containing the smoothed state estimates "s" and the covariances "S"
}
\description{
Implements fixed-interval smoothing for the
Unscented Kalman filter (UKF). The function uses the Rauch
Tung, and Striebel (RTS) smoother for computing the smoothed estimates.
}
\references{
Sarkka, 2008 "Unscented Rauch-Tung-Striebel Smoother"
}
\author{
Stefan Gelissen
}
