% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ydaytime}
\alias{ydaytime}
\title{yday with decimal time}
\usage{
ydaytime(x)
}
\arguments{
\item{x}{as POSIXct vector}
}
\value{
numeric vector of day of the year with decimal time
}
\description{
converts POSIXct to day of year with decimal time.
}
\details{
The first day of the year is treated as zero, i.e. midday on Jan 1st will return 0.5.
This differs from the value provided by the `yday` lubridate  function
}

