% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathy_match}
\alias{bathy_match}
\title{Extract depth for position from GEBCO}
\usage{
bathy_match(lon, lat)
}
\arguments{
\item{lon}{longitude in decimal degrees}

\item{lat}{latitude in decimal degrees}
}
\value{
depth in meters
}
\description{
extract nearest point from GEBCO 2022 at 0.05 degree resolution
}
\examples{
lons = c(0.0001, -15)
lats = c(54.001, 60)
bathy_match(lons, lats)
}
