% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathymap}
\alias{bathymap}
\title{GEBCO Bathymetry base map}
\usage{
bathymap(
  lon = c(-14.996, 8.004),
  lat = c(47, 60),
  margin = 8,
  breaks = T,
  highres = F
)
}
\arguments{
\item{lon}{vector of longitude coordinates for calculating map extent}

\item{lat}{as above for latitude}

\item{margin}{integer (default = 8) indicating fraction of range to use for a margin.}

\item{breaks}{if true (default) depths are binned to <25, 25-50, 50-100, 100-200 and >200m bins}

\item{highres}{default = False, if true fetch the full half degree GEBCO 2019 data}
}
\value{
ggplot
}
\description{
Basemap using GEBCO 2019 data, if you want something more technical, perhaps try the marmap package.
}
\references{
GEBCO data from GEBCO 2019

Coastlines from rworldmap
}
