% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathymap}
\alias{bathymap}
\title{GEBCO Bathymetry base map}
\usage{
bathymap(
  lon = c(-15, 8),
  lat = c(45, 64.5),
  breaks = T,
  highres = F,
  expand = 0.02
)
}
\arguments{
\item{lon}{vector of longitude coordinates for calculating map extent}

\item{lat}{as above for latitude}

\item{breaks}{if true (default) depths are binned to <25, 25-50, 50-100, 100-200 and >200m bins}

\item{highres}{default = False, if true fetch the full half degree GEBCO 2019 data}

\item{expand}{expansion factor for margins, default = 0.02}
}
\value{
ggplot
}
\description{
Basemap using GEBCO 2019 data, if you want something more technical, perhaps try the marmap package.
}
\references{
GEBCO data from GEBCO 2019

Coastlines from rworldmap
}
