% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathymap}
\alias{bathymap}
\title{GEBCO Bathymetry base map}
\usage{
bathymap(lat = c(47, 60), lon = c(-14.996, 8.004), bathy_file = NA,
  breaks = T)
}
\arguments{
\item{lat}{vector of latitude coordinates for calculating map extent}

\item{lon}{as above for longitude}

\item{bathy_file}{optional bathymetry raster file}

\item{breaks}{if true (default) depths are binned to <25, 25-50, 50-100, 100-200 and >200m bins}
}
\value{
ggplot
}
\description{
Simple basemap for using GEBCO data, if you want something more technical try marmap.
}
\references{
GEBCO data from GEBCO 2014

Coastlines from rworldmap
}

