% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{bathymap}
\alias{bathymap}
\title{GEBCO Bathymetry base map}
\usage{
bathymap(
  lon = c(-14, 9),
  lat = c(46, 62),
  breaks = T,
  highres = F,
  expand = 0.02
)
}
\arguments{
\item{lon}{vector of longitude coordinates for calculating map extent}

\item{lat}{as above for latitude}

\item{breaks}{if true (default) depths are binned to <25, 25-50, 50-100, 100-200 and >200m bins}

\item{highres}{default = False, if True fetch the full 0.00417 degree GEBCO 2022 data, otherwise uses 0.05 degree binned data}

\item{expand}{expansion factor for margins, default = 0.02}
}
\value{
ggplot
}
\description{
Basemap using GEBCO 2022 data, if you want something more technical, perhaps try the marmap package.
}
\references{
GEBCO data from GEBCO 2022

Coastlines from rworldmap
}
