% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{convert_latlong}
\alias{convert_latlong}
\title{Convert degrees + decimal minutes to decimal degrees}
\usage{
convert_latlong(degrees, decimal_minutes, polarity = NA)
}
\arguments{
\item{degrees}{numeric vector of whole degrees, if negative will convert to W or S.}

\item{decimal_minutes}{numeric vector of decimal minutes}

\item{polarity}{optional "E, W, N, S"}
}
\value{
decimal degrees
}
\description{
Convert degrees + decimal minutes to decimal degrees
}

