% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.errorcode}
\alias{ferrybox.errorcode}
\title{Ferrybox error codes}
\usage{
ferrybox.errorcode(x, collapse_vector = T)
}
\arguments{
\item{x}{error code}

\item{collapse_vector}{if true single element vector will be returned with ";" seperators}
}
\value{
codes used in composite
}
\description{
Flags are:
0 = no flag, good data
1 = Bad or missing data (general flag)
2 = flow rate too high
4 = flow rate too low
8 = timeout
16 = standby, fb loop not running
32 = wrong channel
64 = value during cleaning
128 = value during standby
256 = value during empty
512 = value during general error
1024 = undefined
2048 = value is simulated
}

