% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.errorcode}
\alias{ferrybox.errorcode}
\title{Ferrybox error codes}
\usage{
ferrybox.errorcode(x, collapse_vector = T)
}
\arguments{
\item{x}{error code}

\item{collapse_vector}{if true single element vector will be returned with ";" seperators}
}
\value{
codes used in composite
}
\description{
Ferrybox error codes
}

