% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferrybox.R
\name{ferrybox.speed}
\alias{ferrybox.speed}
\title{calculate speed from GPS}
\usage{
ferrybox.speed(dateTime, lat, lon, threshold = 65, unit = "knots")
}
\arguments{
\item{dateTime}{in POSIXct}

\item{lat}{in decimal degrees}

\item{lon}{in decimal degrees}

\item{threshold}{maximum time interval between gps points (seconds)}

\item{unit}{"knots" (default) or "ms" for meters per second}
}
\value{
vector of speeds
}
\description{
Make sure your data.frame or table is ordered by dateTime.
}
