% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findMLD}
\alias{findMLD}
\title{Find mixed layer depth}
\usage{
findMLD(
  depth,
  density,
  threshold = 0.125,
  ref_depth = NA,
  surface = T,
  band = F
)
}
\arguments{
\item{depth}{numeric vector of depth}

\item{density}{vector of density}

\item{threshold}{numeric density threshold, default is 0.125}

\item{ref_depth}{reference depth for threshold, default is shallowest depth available}

\item{surface}{default is true, set to false for bottom mixed layer threshold (base of gradient)}

\item{band}{default is false, if true function returns paired vector of interval which meets threshold}
}
\value{
mld
}
\description{
simple threshold technique
}
\details{
returns max depth if threshold not met
}
