% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findMLD}
\alias{findMLD}
\title{Find mixed layer depth}
\usage{
findMLD(depth, density, threshold = 0.125)
}
\arguments{
\item{threshold}{}
}
\value{
mld
}
\description{
simple threshold technique, mld where p +/- 0.125 of surface p
}
\details{
returns max depth is no stratification found
}

