% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findMLD}
\alias{findMLD}
\title{Find mixed layer depth}
\usage{
findMLD(depth, density, threshold = 0.125, surface = T)
}
\arguments{
\item{surface}{default is true, set to false for bottom mixed layer threshold (base of gradient)}
}
\value{
mld
}
\description{
simple threshold technique, mld where p +/- 0.125 of surface p
}
\details{
returns max depth if threshold not met
}

