% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{findMLD}
\alias{findMLD}
\title{Find mixed layer depth}
\usage{
findMLD(z, y, threshold = 0.125, ref_z = NA, surface = T, band = F)
}
\arguments{
\item{z}{numeric vector of z (either pressure or depth)}

\item{y}{vector of MLD indicating variable, typically temperature, or density}

\item{threshold}{numeric y threshold, default is 0.125 (assuming y)}

\item{ref_z}{reference z for threshold, default is shallowest z available}

\item{surface}{default is true, set to false for bottom mixed layer threshold (base of gradient)}

\item{band}{default is false, if true function returns paired vector of interval which meets threshold}
}
\value{
mld
}
\description{
simple threshold technique
}
\details{
returns max z if threshold not met
}
\examples{
x = data.table(pressure = 1:100,
               density = approx(c(1, 40, 50, 80, 100),
                                c(1.125, 1.125, 1.293, 1.308, 1.308),
                                xout = 1:100)$y)

findMLD(x$pressure, x$density)
}
