% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggplot.lm}
\alias{ggplot.lm}
\title{ggplot lm}
\usage{
\method{ggplot}{lm}(fit)
}
\arguments{
\item{fit}{a lm fit object}
}
\value{
ggplot2 dot plot with equation at top
}
\description{
quickly plot a lm like everyone wants from excel
}
