% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lagged_flag}
\alias{lagged_flag}
\title{lagged flag}
\usage{
lagged_flag(dateTime, init, lag = 300)
}
\arguments{
\item{dateTime}{a vector of times which are to be flagged, this does not need to be POSIXct}

\item{lag}{integer period to add to times, should be in the same units i.e. seconds for POSIXct}
}
\value{
logical vector of same length as dateTime, True denotes flagged
}
\description{
function designed to flag data based on specified start times
This is mostly used for flagging data for x minutes after an event
}
