% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.cruiselist}
\alias{lims.cruiselist}
\title{Fetch list of cruises from LIMS}
\usage{
lims.cruiselist(yr = "ALL", db_name = "lims")
}
\arguments{
\item{yr}{}

\item{db_name}{}
}
\value{
vector of cruise ids
}
\description{
Fetch list of cruises from LIMS
}

