% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.fetch}
\alias{lims.fetch}
\title{Fetch nutrients data from LIMS (post 2013)}
\usage{
lims.fetch(parameters = "SAL", cruise = NA, db_name = "lims")
}
\arguments{
\item{parameters}{}

\item{cruise}{}

\item{db_name}{}
}
\value{
data frame (data.table) of extracted nutrients data
}
\description{
Fetch nutrients data from LIMS (post 2013)
}
\details{
parameter codes:
Chlorophyll, TOXN, O2, SLD, SAL
}

