% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.fetch}
\alias{lims.fetch}
\title{Fetch nutrients data from LIMS (post 2013)}
\usage{
lims.fetch(parameters = c("SAL", "CHLOROPHYLL", "SLD", "TOXN", "O2"),
  cruise = NA, db_name = "lims")
}
\value{
data frame (data.table) of extracted nutrients data
}
\description{
Fetch nutrients data from LIMS (post 2013)
}
\details{
parameter codes:
Chlorophyll, TOXN, O2, SLD, SAL
}

