% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims.R
\name{lims.fetch}
\alias{lims.fetch}
\title{Fetch nutrients data from LIMS}
\usage{
lims.fetch(
  parameters = c("SAL", "CHLOROPHYLL", "SPM", "TOXN", "SIO4", "NH4", "PHAEOP", "NO2",
    "PO4", "DO2"),
  cruise = NA,
  after = NA,
  before = NA,
  convert = T,
  area = NA,
  db_name = "lims"
)
}
\arguments{
\item{parameters}{character string of LIMS parameter names}

\item{cruise}{optional character string of cruise name}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{convert}{convert fields to numeric? default = True}

\item{area}{optional vector consisting of 4 elements, max Latitude, max Longitude, min Latitude, min Longitude e.g. c(53, -2.5, 52, -4)}

\item{db_name}{character string matching ODBC data source name, defaults to 'lims'}
}
\value{
data frame (data.table) of extracted nutrients data
}
\description{
Fetch nutrients data from LIMS
}
\details{
parameter codes:
Chlorophyll, TOXN, O2, SLD, SAL

note "value" returned as character string to allow for <LOD values
}
