% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{markday}
\alias{markday}
\title{Calculate day or night}
\usage{
markday(dateTime, lat, lon)
}
\arguments{
\item{dateTime}{POSIXct datetime}

\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}
}
\value{
vector of "day" or "night" character values
}
\description{
Calculates if a given datetime is during day or night based on sun angle.
}
