% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{optode.phaseCalc}
\alias{optode.phaseCalc}
\title{Optode phase}
\usage{
optode.phaseCalc(phase, Temp, coefs)
}
\arguments{
\item{Temp}{vector of temperature (C)}

\item{coefs}{list (or data.frame) consisting of C0..C6 coefs, see examples}

\item{DPhase}{vector of phase values}
}
\value{
oxygen concentration in mmol m-3
}
\description{
Uses suppled optode calibration coeffients to calculate oxygen concentration from Dphase/Calphase and temperature.
}
\examples{

# for SVU multipoint calibrated optodes:
SVU_coef = list(batch="1517M", coef="SVU", C0=0.002757413, C1=0.000115132, C2=2.34E-06, C3=234.4436, C4=-0.3752192, C5=-45.60156, C6=4.630104)
optode.phaseCalc(45, 10, coefs=SVU_coef)

# For standard 4330, 4831, 4835 optodes:
coefs = list(batch = "4807E", coef = "mk2",
  FoilCoefA = c(-2.988314E-06, -6.137785E-06, 1.684659E-03, -1.857173E-01, 6.784399E-04, -5.597908E-07, 1.040158E+01,
                -5.986907E-02, 1.360425E-04, -4.776977E-07, -3.032937E+02, 2.530496E+00, -1.267045E-02, 1.040454E-04),
  FoilCoefB = c(-3.560390E-07, 3.816713E+03, -4.475507E+01, 4.386164E-01, -7.146342E-03, 8.906236E-05, -6.343012E-07,
                0.000000E+00, 0.000000E+00, 0.000000E+00, 0.000000E+00, 0.000000E+00, 0.000000E+00, 0.000000E+00),
  FoilPolyDegT = c(1, 0, 0, 0., 1, 2, 0, 1, 2, 3, 0, 1, 2, 3, 4, 0, 1, 2, 3, 4, 5, 0, 0, 0, 0, 0, 0, 0),
  FoilPolyDegO = c(4, 5, 4, 3, 3, 3, 2, 2, 2, 2, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
optode.phaseCalc(30, 10, coefs=coefs)

# For standard 3835 optodes:
coefs = list(
  batch = 1707,
  coef = "mk1",
  C0 = c(5326.5, -192.117, 4.14357, -0.0378695),
  C1 = c(-292.068, 9.71993, -0.214295, 0.00200778),
  C2 = c(6.47595, -0.19808, 0.0044994, -0.0000431),
  C3 = c(-0.0669288, 0.00188066, -0.0000442, 0.000000428),
  C4 = c(0.000265042, -0.00000683, 0.000000167, -1.62E-09))
optode.phaseCalc(30, 10, coefs=coefs)
}
