% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxygen.R
\name{oxygen.air_conc}
\alias{oxygen.air_conc}
\title{Oxygen in-air concentration}
\usage{
oxygen.air_conc(TEMP, SAL, AIRPRS, RH = NA, DTEMP = NA, return_conc = T)
}
\arguments{
\item{TEMP}{air temperature in Celsius}

\item{SAL}{salinity}

\item{AIRPRS}{air pressure in mbar/hPa}

\item{RH}{relative humidity in \%}

\item{DTEMP}{dewpoint temperature in Celsius}

\item{return_conc}{if True (default) return concentration, otherwise returns partial pressure in hPa}
}
\value{
oxygen concentration in mmol m-3
}
\description{
Calculates in-air concentration of oxygen, given temperature, pressure and relative humidity.
If only given dewpoint will use Sargent (1980) approximation to derive relative humidity
}
\details{
Bittig2018
}
\keyword{oxygen}
