% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{PAR_from_Voltage}
\alias{PAR_from_Voltage}
\title{Calculate Cefas LiCor PAR from voltage}
\usage{
PAR_from_Voltage(x, factor, offset)
}
\arguments{
\item{x}{vector of output voltage (LICOR)}

\item{factor}{licor factor in volts}

\item{offset}{licor offset in volts}
}
\value{
numeric PAR in uE m-2 s-1
}
\description{
Calculates PAR from Licor Li-198 sensors
 fitted with Cefas inline low-light amplifier.
 Refer to calibration documents for factor and offset.
}

