% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.binning}
\alias{profiler.binning}
\title{ESM2 profiler depth binning}
\usage{
profiler.binning(x, bin_depth = 0.5, method = floor, use_cast = "UP",
  return_bin = "all")
}
\arguments{
\item{x}{data.frame matching output from fetch.profiler}

\item{bin_depth}{numeric vector determining depth binning interval, default is 0.5}

\item{method}{function for binning, default is floor, ceil and round also work}

\item{use_cast}{character string matching matching cast required, options are 'UP'}

\item{return_bin}{numeric vector, only these bin depths will be returned}
}
\value{
character vector of Cruise Id's
}
\description{
Binns profiler data into depth bins
}
\details{
TODO
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}

