% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.binning}
\alias{profiler.binning}
\title{ESM2 profiler depth binning}
\usage{
profiler.binning(
  x,
  bin_height = 1,
  method = round,
  use_cast = "UP",
  return_bin = "all",
  O2_trim = F
)
}
\arguments{
\item{x}{data.frame matching output from fetch.profiler}

\item{bin_height}{numeric vector determining depth binning interval, default is 0.5}

\item{method}{function for binning, default is round, ceiling and floor also work}

\item{use_cast}{character string matching matching cast required, options are UP, DOWN and ALL}

\item{return_bin}{numeric vector, only these bin depths will be returned}

\item{O2_trim}{if true the first measurement in each bin is dropped, this is to give the sensor time to respond}
}
\value{
character vector of Cruise Id's
}
\description{
Bins profiler data into depth bins
}
\details{
This function uses the standard "boxcar" approch to binning CTD profiles.
profiles can be split to use the average value at each depth from both up and down casts (the "ALL" option),
or the binning can be restricted to just the down or up casts. DOWN is usually best when the profiler has been deployed correctly,
UP is useful for comparison to niskin bottle firings.
Note that startTime and endTime corrispond to the time at the start or end of the cast.
e.g. if you select UP ast the use_cast option startTime will be the time at the start of the UP cast.
The UP cast is identified as being the time after the deepest part of the profile.
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}
