% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.binning}
\alias{profiler.binning}
\title{ESM2 profiler depth binning}
\usage{
profiler.binning(x, bin_height = 1, method = round, use_cast = "UP",
  return_bin = "all")
}
\arguments{
\item{x}{data.frame matching output from fetch.profiler}

\item{bin_height}{numeric vector determining depth binning interval, default is 0.5}

\item{method}{function for binning, default is round, ceiling and floor also work}

\item{use_cast}{character string matching matching cast required, options are UP, DOWN and ALL}

\item{return_bin}{numeric vector, only these bin depths will be returned}
}
\value{
character vector of Cruise Id's
}
\description{
Binns profiler data into depth bins
}
\details{
TODO, default boxcar
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}

