% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.eu_depth}
\alias{profiler.eu_depth}
\title{estimate euphotic depth from PAR profile}
\usage{
profiler.eu_depth(depth, par, surf_par)
}
\arguments{
\item{depth}{vector of depth values}

\item{par}{vector of par values}

\item{surf_par}{surface par}
}
\value{
euphotic depth
}
\description{
Euphotic depth defined as 1% of surface
}
