% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.fetch}
\alias{profiler.fetch}
\title{ESM2 profiler data query tool}
\usage{
profiler.fetch(cruiseID = NA, profiler = NA, after = NA, before = NA,
  area = NA, parameters = c("TEMP", "SAL", "FTU", "O2CONC", "PAR",
  "FLUORS"), RQ0 = TRUE, ct_temp_only = TRUE, min_QA_reached = TRUE,
  db_name = "smartbuoydblive")
}
\arguments{
\item{cruiseID}{optional cruise ID string, e.g. "CEND_02_14", if provided only data from this cruise will be returned.}

\item{profiler}{optional profiler name string, e.g. "PR009", if provided only data from this profiler will be returned.}

\item{after}{optional date string, if provided only data after this date will be returned, assumes UTC e.g. "2014-08-10"}

\item{before}{optional date string, if provided only data before this date will be returned, assumes UTC e.g. "2014-12-09"}

\item{area}{optional area vector consisting of 4 elements, minimum Latitude, minimum Longitude, maximum Latitude, maximum Longitude e.g. c(52, -3.5, 53, -4)}

\item{parameters}{vector of parameter code names, defaults to c('TEMP', 'SAL', 'FTU', 'O2CONC', 'PAR', 'FLUORS')}

\item{RQ0}{boolean, if True only data where result quality = 0 is returned, i.e. good data. default is True}

\item{ct_temp_only}{boolean, if True all non-FSI temperature data is discarded, default is False.}

\item{min_QA_reached}{boolean, if True only data which has passed required QA level is returned, always used with QA0 = True.}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.frame with returned data in "long" format or error string if no data returned
}
\description{
Fetches CTD data from ESM2 profiler records
}
\details{
This function querys the Smartbuoy database and returns ESM2 profiler data matching the provided critiera.
the v_CtdProfile_AllData table is used, as such private data will not be available to this function.
}
\keyword{ctd}
\keyword{esm2}
\keyword{profiler}
\keyword{query}

