% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiler.R
\name{profiler.header}
\alias{profiler.header}
\title{ESM2 profiler header queryer}
\usage{
profiler.header(yr = "ALL", db_name = "smartbuoydblive")
}
\arguments{
\item{yr}{integer year defaults to "ALL"}

\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
data.frame of headers
}
\description{
ESM2 profiler header queryer
}

