% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{read.ecmwf}
\alias{read.ecmwf}
\title{Read and convert ECMWF netcdf to data.table}
\usage{
read.ecmwf(file)
}
\arguments{
\item{file}{}
}
\value{
data.table containing all netcdf variables in long format
}
\description{
Read and convert ECMWF netcdf to data.table
}
