% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.ego}
\alias{read.ego}
\title{Read BODC ego glider files (e.g. slocum)}
\usage{
read.ego(
  ncfile,
  vars = c("DOXY", "PRES", "TEMP", "SAL", "CNDC", "CHLA", "BBP700")
)
}
\arguments{
\item{ncfile}{ego netcdf file}

\item{vars}{character vector of variables to extract, e.g. c("DOXY", "PRES")}
}
\value{
data.table containing extracted fields
}
\description{
Reads EGO glider time-series data
Currently only tested with v1.2
}
