% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.10min}
\alias{read.ferrybox.10min}
\title{Ferrybox 10min file data reader}
\usage{
read.ferrybox.10min(folder, recursive = F, print_file = T)
}
\arguments{
\item{folder}{folder containing 10min files}

\item{recursive}{if true look in subfolders}

\item{print_file}{if true filenames are printed as processed}
}
\value{
data.frame (data.table) of processed 10min files
}
\description{
reads JENA-4H 10min files
}
\details{
TODO
}
\keyword{10minfile}
\keyword{ferrybox}

