% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.10min}
\alias{read.ferrybox.10min}
\title{Ferrybox 10min file data reader}
\usage{
read.ferrybox.10min(folder, save = T, recursive = F)
}
\arguments{
\item{folder}{folder containing 10min files}

\item{save}{if true tool will write to 10mindat.rdata file while reading}

\item{recursive}{if true look in subfolders}
}
\value{
data.frame (data.table) of processed 10min files
}
\description{
reads JENA-4H 10min files
}
\details{
TODO
}
\keyword{10minfile}
\keyword{ferrybox}

