% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.ferrybox.devdata}
\alias{read.ferrybox.devdata}
\title{Ferrybox device data reader}
\usage{
read.ferrybox.devdata(devdata_folder, pivot = F)
}
\arguments{
\item{devdata_folder}{character string indicating path to folder containing device data files}

\item{pivot}{optional boolian indicating if returned table should be recast into 'wide' format}
}
\value{
data.frame
}
\description{
reads ferrybox device data files
}
\details{
TODO
Ferrybox quality codes:
0 = good, 1 = no data, 2 = over value, 4 = under value, 8 = sensor timeout
64 = clean cycle, 128 = standby, 256 = empty, 512 = error, 1024 = undefined, 2048 = simulation mode
sometimes \0 characters appear in data, to sanitise run sed from a bash terminal
}
\keyword{ferrybox}
