% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hardware.R
\name{read.liveAquire}
\alias{read.liveAquire}
\title{ESM2 live aquire reader}
\usage{
read.liveAquire(file = dlgOpen(title = "Open LiveAquire file...")$res)
}
\arguments{
\item{file}{location of live aquire file, leave blank to use dialog box}
}
\value{
a list containing a data frame and ggplot
}
\description{
reads live aquire files and plots output
}
\details{
This function reads text files generated by the SmartBuoy interface live aquire.
}
\keyword{esm2}
\keyword{smartbuoy}

