% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.liveAquireAnalog}
\alias{read.liveAquireAnalog}
\title{read analog live aquire data}
\usage{
read.liveAquireAnalog(x, channels_table = data.frame(channel = c(0, 1, 6),
  type = c("OBS", "FLU", "OBS"), slope = c(1.22, 1.22, 1.22), offset = c(0, 0,
  0)))
}
\arguments{
\item{x}{one or more files captured from liveaquire}

\item{channels_table}{}
}
\value{
processed data table
}
\description{
read analog live aquire data
}

