% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_readers.R
\name{read.NMEA_GGA}
\alias{read.NMEA_GGA}
\title{read and parse NMEA GGA log file}
\usage{
read.NMEA_GGA(file, date_origin = "2000-01-01", only_good = T)
}
\arguments{
\item{file}{filename}

\item{date_origin}{character string containing the date origin in "YYYY-MM-DD"}

\item{only_good}{if True (default) discard lines with bad GPS fixes}
}
\value{
data.table containing, gpsTime, lat, lon, number of satelites, HDOP (meters) and altitude (meters)
}
\description{
Tool will parse a NMEA log file, and decode each line with a "$xxGGA" string
Note that GGA does not contain the date, only UTC time, so a date origin must be supplied for each file.
If your log file spans multiple days, make sure you handle the roll-over at midnight!
}
\details{
This tool also ignores any prefix to the $xxGGA string start, so preceding Teraterm timestamps are ignored.
}
\examples{
read.NMEA_GGA("teraterm.log", "2020-01-01")
}
