% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{read.seaglider_log}
\alias{read.seaglider_log}
\title{read seaglider logfiles, extracts GPS, altimetry, battery data, and targets}
\usage{
read.seaglider_log(glider_folder, echo = F)
}
\arguments{
\item{echo}{if true print name of log file}
}
\value{
data.table containing processed logfiles
}
\description{
dateTime is from GPS time
}
