% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaglider.R
\name{seaglider.cond}
\alias{seaglider.cond}
\title{Calculate seaglider conductivity}
\usage{
seaglider.cond(condFreq, temp, pressure, calib_file)
}
\arguments{
\item{condFreq}{vector of seaglider SBE temperature frequency}

\item{temp}{temperature as calculated with `seaglider.temp`}

\item{pressure}{in dbar as calculated with `seaglider.pressure`}

\item{calib_file}{path to seaglider calibration.mat file}
}
\value{
vector of calibrated conductivity in S/m
}
\description{
Direct method, uses temperature and depth at cond sample time.
No correction for flow speed or thermal inertia applied
}
