% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.positions}
\alias{smartbuoy.positions}
\title{SmartBuoy positions}
\usage{
smartbuoy.positions(db_name = "smartbuoydblive", group = T)
}
\arguments{
\item{db_name}{character string matching ODBC data source name, defaults to 'smartbuoydblive'}

\item{group}{if True (default) aggregates position by deployment group}
}
\value{
data.table of positions
}
\description{
Fetches median positions of deployments
}
\keyword{SmartBuoy}

