% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.telemetry_position}
\alias{smartbuoy.telemetry_position}
\title{Query smartbuoy deployment position from telemetry messages}
\usage{
smartbuoy.telemetry_position(deployment, db_name = "smartbuoydblive")
}
\arguments{
\item{deployment}{}
}
\value{
data.table of lat-long from telemetry messages
}
\description{
Query smartbuoy deployment position from telemetry messages
}

