% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.telemetry_position}
\alias{smartbuoy.telemetry_position}
\title{Query smartbuoy deployment position from telemetry messages}
\usage{
smartbuoy.telemetry_position(
  deployment = NA,
  deployment_group = NA,
  db_name = "smartbuoydblive"
)
}
\arguments{
\item{deployment}{provide if you want a specific buoy deployment e.g. "DOWSING/040"}

\item{deployment_group}{provide if you want everything for a site e.g. "WESTGAB2"}
}
\value{
data.table of lat-long from telemetry messages
}
\description{
Query smartbuoy deployment position from telemetry messages
}
