% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartbuoy.R
\name{smartbuoy.TS}
\alias{smartbuoy.TS}
\title{SmartBuoy T/S plots}
\usage{
smartbuoy.TS(deployment, db_name = "smartbuoydblive")
}
\arguments{
\item{deployment}{character string matching SmartBuoy deployment (case sensitive)}

\item{db_name}{optional character string matching ODBC data source name, defaults to 'smartbuoydblive'}
}
\value{
ggplot object
}
\description{
Draws a T/S plot for a SmartBuoy Deployment
}
\details{
TODO
}
\keyword{esm2}

