% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_glider}
\alias{theme_glider}
\title{Glider ggplot2 theme}
\usage{
theme_glider()
}
\value{
ggplot2 theme
}
\description{
based on theme_bw, moves legend to the bottom, makes it wide, useful for matlab style glider plots
}
